% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_thresholds_transition_risk.R
\name{add_thresholds_transition_risk}
\alias{add_thresholds_transition_risk}
\title{Adds low and high thresholds for transition risk score}
\usage{
add_thresholds_transition_risk(
  emissions_profile_products,
  all_uuids_scenario_sectors,
  scenarios
)
}
\arguments{
\item{emissions_profile_products}{A dataframe}

\item{all_uuids_scenario_sectors}{A dataframe}

\item{scenarios}{A dataframe}
}
\value{
A dataframe.
}
\description{
Adds low and high thresholds for transition risk score
}
\examples{
library(tiltToyData)
library(readr)
library(dplyr)
options(readr.show_col_types = FALSE)

emissions_profile_products <- read_csv(toy_emissions_profile_products_ecoinvent())
all_uuids_scenario_sectors <- read_csv(toy_sector_profile_companies()) |>
  select(-c("companies_id", "company_name", "clustered")) |>
  distinct()
scenarios <- read_csv(toy_sector_profile_any_scenarios())

output <- add_thresholds_transition_risk(
  emissions_profile_products,
  all_uuids_scenario_sectors,
  scenarios
)
output
}
\keyword{internal}
