% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_aka.R
\name{aka}
\alias{aka}
\title{Access a column via an internal alias it's 'also known as'}
\usage{
aka(x, dictionary = example_dictionary())
}
\arguments{
\item{x}{A character giving the internal 'also known as' of a column name.}

\item{dictionary}{A dataframe like \code{\link[=example_dictionary]{example_dictionary()}}.}
}
\value{
A character.
}
\description{
Accessing columns this way makes the code more maintainable by avoiding
column names that are likely to change.
}
\examples{
aka("id")
}
\keyword{internal}
