% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{categorize_risk}
\alias{categorize_risk}
\title{Categorize a variable into risk categories using low and high thresholds}
\usage{
categorize_risk(x, low_threshold, high_threshold, ...)
}
\description{
Categorize a variable into risk categories using low and high thresholds
}
\keyword{internal}
