% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{companies}
\alias{companies}
\title{XCTR datasets}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9 rows and 4 columns.
}
\usage{
companies
}
\description{
XCTR datasets
}
\examples{
companies
}
\author{
Kalash Singhal.
}
\concept{XCTR datasets}
\keyword{datasets}
