% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_profile_ranking.R
\name{compute_profile_ranking}
\alias{compute_profile_ranking}
\alias{spa_compute_profile_ranking}
\alias{epa_compute_profile_ranking}
\title{Compute profile ranking}
\usage{
spa_compute_profile_ranking(data, scenarios)

epa_compute_profile_ranking(data)
}
\description{
Compute profile ranking
}
\keyword{internal}
