% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_thresholds_transition_risk.R
\name{create_tr_benchmarks_tr_score}
\alias{create_tr_benchmarks_tr_score}
\title{Calulate \code{transition_risk_score} and \code{benchmark_tr_score} columns}
\usage{
create_tr_benchmarks_tr_score(data, profile_ranking, reduction_targets)
}
\arguments{
\item{data}{Dataframe.}

\item{profile_ranking}{Dataframe column.}

\item{reduction_targets}{Dataframe column.}
}
\description{
Calulate \code{transition_risk_score} and \code{benchmark_tr_score} columns
}
\keyword{internal}
