% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dir.R
\name{data_dir}
\alias{data_dir}
\title{Path to user data directories}
\usage{
data_dir(version = NULL, ...)
}
\arguments{
\item{version}{is an optional version path element to append to the
path. You might want to use this if you want multiple versions
of your app to be able to run independently. If used, this
would typically be \code{"<major>.<minor>"}. Only applied when appname
is not NULL.}

\item{...}{
  Arguments passed on to \code{\link[rappdirs:user_data_dir]{rappdirs::user_data_dir}}
  \describe{
    \item{\code{appname}}{is the name of application. If NULL, just the system
directory is returned.}
    \item{\code{appauthor}}{(only required and used on Windows) is the name of the
appauthor or distributing body for this application. Typically
it is the owning company name. This falls back to appname.}
    \item{\code{roaming}}{(logical, default \code{FALSE}) can be set \code{TRUE} to
use the Windows roaming appdata directory. That means that for users on
a Windows network setup for roaming profiles, this user data will be
sync'd on login. See
\url{https://docs.microsoft.com/en-us/previous-versions/windows/it-pro/windows-vista/cc766489(v=ws.10)}
for a discussion of issues.}
    \item{\code{expand}}{If TRUE (the default) will expand the \code{R_LIBS} specifiers with their equivalents.
See \code{\link[=R_LIBS]{R_LIBS()}} for list of all possibly specifiers.}
    \item{\code{os}}{Operating system whose conventions are used to construct the
requested directory. Possible values are "win", "mac", "unix". If \code{NULL}
(the default) then the current OS will be used.}
  }}
}
\value{
A NULL.
}
\description{
Path to user data directories
}
\examples{
data_dir()
data_dir(version = "x.y.z")
}
