% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{istr_at_product_level}
\alias{pstr_at_product_level}
\alias{xctr_at_product_level}
\alias{istr_at_company_level}
\alias{pstr_at_company_level}
\alias{xctr_at_company_level}
\alias{xstr_pivot_type_sector_subsector}
\alias{xstr_prepare_scenario}
\alias{xstr_prune_companies}
\alias{xstr_polish_output_at_company_level}
\title{Deprecated}
\usage{
istr_at_product_level(
  companies,
  scenarios,
  inputs,
  low_threshold = ifelse(scenarios$year == 2030, 1/9, 1/3),
  high_threshold = ifelse(scenarios$year == 2030, 2/9, 2/3)
)

pstr_at_product_level(
  companies,
  scenarios,
  low_threshold = ifelse(scenarios$year == 2030, 1/9, 1/3),
  high_threshold = ifelse(scenarios$year == 2030, 2/9, 2/3)
)

xctr_at_product_level(
  companies,
  co2,
  low_threshold = 1/3,
  high_threshold = 2/3
)

istr_at_company_level(data)

pstr_at_company_level(data)

xctr_at_company_level(data)

xstr_pivot_type_sector_subsector(data)

xstr_prepare_scenario(scenarios)

xstr_prune_companies(data)

xstr_polish_output_at_company_level(data)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

All functions of the form \code{f_at_product_level()} and \code{f_at_company_level()}
are now deprecated in favor of higher-level wrappers (see
\code{?tiltIndicator::rename}).
}
\keyword{internal}
