% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{document_default_value}
\alias{document_default_value}
\title{Document the default return value}
\usage{
document_default_value()
}
\description{
Document the default return value
}
\examples{
document_default_value()
}
\keyword{internal}
