% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\docType{data}
\name{e}
\alias{e}
\title{Avoid R CMD Check warning about undocumented data sets}
\format{
An object of class \code{environment} of length 8.
}
\usage{
e
}
\description{
R CMD check wants documentation for anything defined in this file. I stick
all objects in \code{e} so that I can document that single \code{e} object as internal.
}
\keyword{internal}
