% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoinvent_activities_demo.R
\docType{data}
\name{ecoinvent_activities_demo}
\alias{ecoinvent_activities_demo}
\title{A public sample dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 20 rows and 16 columns.
}
\source{
Tilman via https://github.com/2DegreesInvesting/tiltData/issues/200
}
\usage{
ecoinvent_activities_demo
}
\description{
Public based on a meeting on 2023-03-22
}
\examples{
ecoinvent_activities_demo
}
\keyword{datasets}
