% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissions_profile.R
\name{emissions_profile}
\alias{emissions_profile}
\title{Calculate the indicator "emissions profile"}
\usage{
emissions_profile(companies, co2, low_threshold = 1/3, high_threshold = 2/3)
}
\arguments{
\item{companies, co2}{A dataframe like the dataset with a matching name in tiltToyData (see \href{https://2degreesinvesting.github.io/tiltToyData/reference/index.html}{Reference}).}

\item{low_threshold}{A numeric value to segment low and medium transition
risk products.}

\item{high_threshold}{A numeric value to segment medium and high transition
risk products.}
}
\value{
A data frame with the column \code{companies_id}, and the nested columns\code{product} and \code{company} holding the outputs at product and company level. Unnesting \code{product} yields a data frame with at least columns \code{companies_id}, \code{grouped_by}, \code{risk_category}. Unnesting \code{company} yields a data frame with at least columns \code{companies_id}, \code{grouped_by}, \code{risk_category}, \code{value}. Any column in the input datasets ending with \verb{*rowid} is also passed as is to the output at product level. The exception is any column named exactly \code{rowid}-- which is a reserved name and throws an error. Note this feature makes no sense at company level because potentially multiple rows in the input datasets are summarized into a single row in the output at company level.
}
\description{
The "emissions profile" measures the transition risk on the product-level of a
company. The indicator is expressed in the percentage of all products that are
at high risk, medium risk or low risk due to the products' relative greenhouse
gas (GHG) emissions. The assessment is first performed on a product-level and
can then be aggregated to the company-level.

The "emissions profile" measures the relative GHG emissions per product. As a
default option, each product is compared to the GHG emissions of every other
product. Products with higher GHG emissions are considered to have a higher
emission profile and hence also face a higher climate risk.

After identifying the GHG emissions for each product, the products are ranked
according to their carbon footprint. The ranking method is explained in more
detail in the
\href{https://2degreesinvesting.github.io/tiltIndicator/articles/tiltIndicator.html#thresholds}{Thresholds}
section. In short, products with relatively low GHG emissions are assigned to
the category "low", products with relatively high GHG emissions to the category
"high" and the products in between to the category "medium".

Note that for the ranking and categorization, we apply different benchmarks.
Products are compared to (i) all other products, to (ii) products within the
same tilt sector, to (iii) products within the same ISIC 4 digit sector, to (iv)
products measured in the same unit, and a combination of (iv) with either (i),
(ii) or (iii).

For the company-level results, we aggregate all products from the same category
and set them in relation to all products that the company produces. We derive
the "emissions profile", which is expressed as the share of products per
category for each company.

Finally, the indicator provides the share of products with "low", "medium", or
"high" relative GHG emissions per company.

The output of this indicator contains the following:
\itemize{
\item A column indicating the benchmark to which a product is compared.
\item A column indicating whether the product has "low", "medium" or "high" relative
GHG emissions.
\item A column indicating the share of the products per category and benchmark".
}
}
\examples{
library(tiltIndicator)
library(tiltToyData)
library(readr)
options(readr.show_col_types = FALSE)

companies <- read_csv(toy_emissions_profile_any_companies())
products <- read_csv(toy_emissions_profile_products_ecoinvent())

both <- emissions_profile(companies, products)
both

both |> unnest_product()

both |> unnest_company()
}
\seealso{
Other main functions: 
\code{\link{emissions_profile_upstream}()},
\code{\link{sector_profile_upstream}()},
\code{\link{sector_profile}()}
}
\concept{main functions}
