% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/emissions_profile_any_add_values_to_categorize.R
\name{emissions_profile_any_add_values_to_categorize}
\alias{emissions_profile_any_add_values_to_categorize}
\title{Add \code{values_to_categorize} FIXME find a better name}
\usage{
emissions_profile_any_add_values_to_categorize(data)
}
\arguments{
\item{data}{FIXME point to tiltToyData. A "co2" dataset containing products or upstream-products.}
}
\value{
FIXME add test. The input dataset with the additional column \code{values_to_cagegorize}.
}
\description{
Add \code{values_to_categorize} FIXME find a better name
}
\examples{
# FIXME add examples
}
\seealso{
Other pre-processing helpers: 
\code{\link{sector_profile_any_pivot_type_sector_subsector}()},
\code{\link{sector_profile_any_prepare_scenario}()},
\code{\link{sector_profile_any_prune_companies}()}
}
\concept{pre-processing helpers}
