% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epa_compute_profile_ranking.R
\name{emissions_profile_any_compute_profile_ranking}
\alias{emissions_profile_any_compute_profile_ranking}
\title{Add values to categorize}
\usage{
emissions_profile_any_compute_profile_ranking(data)
}
\arguments{
\item{data}{A "co2-like" data frame -- i.e. containing products or
upstream-products (a.k.a. inputs).}
}
\value{
The input data frame with the additional columns \code{grouped_by} and
\code{profile_ranking} and one row per benchmark per company.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was deprecated because it's internal. Users don't need to
interact with the function itself.
}
\examples{
library(tiltToyData)
library(readr, warn.conflicts = FALSE)
options(readr.show_col_types = FALSE)

companies <- read_csv(toy_emissions_profile_any_companies())

products <- read_csv(toy_emissions_profile_products_ecoinvent())
products |> emissions_profile_any_compute_profile_ranking()

inputs <- read_csv(toy_emissions_profile_upstream_products_ecoinvent())
inputs |> emissions_profile_any_compute_profile_ranking()
}
\keyword{internal}
