% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissions_profile.R,
%   R/emissions_profile_upstream.R
\name{emissions_profile_upstream}
\alias{emissions_profile_upstream}
\title{Calculate the indicator "emissions profile upstream"}
\usage{
emissions_profile_upstream(
  companies,
  co2,
  low_threshold = 1/3,
  high_threshold = 2/3
)
}
\arguments{
\item{companies, co2}{A dataframe like the dataset with a matching name in tiltToyData (see \href{https://2degreesinvesting.github.io/tiltToyData/reference/index.html}{Reference}).}

\item{low_threshold}{A numeric value to segment low and medium transition
risk products.}

\item{high_threshold}{A numeric value to segment medium and high transition
risk products.}
}
\value{
A data frame with the column \code{companies_id}, and the nested columns\code{product} and \code{company} holding the outputs at product and company level. Unnesting \code{product} yields a data frame with at least columns \code{companies_id}, \code{grouped_by}, \code{risk_category}. Unnesting \code{company} yields a data frame with at least columns \code{companies_id}, \code{grouped_by}, \code{risk_category}, \code{value}. Any column in the input datasets ending with \verb{*rowid} is also passed as is to the output at product level. The exception is any column named exactly \code{rowid}-- which is a reserved name and throws an error. Note this feature makes no sense at company level because potentially multiple rows in the input datasets are summarized into a single row in the output at company level.
}
\description{
\if{html}{\out{<div class="sourceCode {r child=extdata_path("child/intro-emissions-profile_upstream.Rmd")}">}}\preformatted{}\if{html}{\out{</div>}}
}
\examples{
library(tiltToyData)
library(readr)
options(readr.show_col_types = FALSE)

companies <- read_csv(toy_emissions_profile_any_companies())
inputs <- read_csv(toy_emissions_profile_upstream_products())

both <- emissions_profile_upstream(companies, inputs)

both |> unnest_product()

both |> unnest_company()
}
\seealso{
Other main functions: 
\code{\link{emissions_profile}()},
\code{\link{sector_profile_upstream}()},
\code{\link{sector_profile}()}
}
\concept{main functions}
