% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epa_at_company_level.R
\name{epa_at_company_level}
\alias{epa_at_company_level}
\title{Calculate indicator's profile output at company level}
\usage{
epa_at_company_level(data)
}
\arguments{
\item{data}{A dataframe}
}
\value{
A dataframe
}
\description{
Calculate indicator's profile output at company level
}
\keyword{internal}
