% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_dictionary.R
\name{example_companies}
\alias{example_companies}
\title{Create example companies}
\usage{
example_companies(...)
}
\arguments{
\item{...}{Passed to \code{\link[=tibble]{tibble()}}.}
}
\value{
A \code{\link[=tibble]{tibble()}}.
}
\description{
Create example companies
}
\examples{
example_companies()
}
\keyword{internal}
