% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data_factory.R
\name{example_data_factory}
\alias{example_data_factory}
\title{Fabricate functions that construct example datasets}
\usage{
example_data_factory(data)
}
\description{
Fabricate functions that construct example datasets
}
\examples{
data <- tibble::tibble(x = 1, y = 1)
example_data <- example_data_factory(data)

example_data()

col <- "x"
example_data(!!col := 1:2)
}
\keyword{internal}
