% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_dictionary.R
\name{example_dictionary}
\alias{example_dictionary}
\title{Dictionary of example data}
\usage{
example_dictionary(remove_id = TRUE)
}
\arguments{
\item{remove_id}{Remove table id's?}
}
\value{
A \code{\link[=tibble]{tibble()}}.
}
\description{
This dataset is created from a noramalized set of tables to allow developers
to add or change example datasets in a very compact and consistent way.
}
\examples{
example_dictionary() |> print(n = Inf)
}
\keyword{internal}
