% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_raw_data.R
\name{example_raw_companies}
\alias{example_raw_companies}
\alias{example_raw_weo}
\alias{example_raw_ipr}
\title{Example raw datasets}
\usage{
example_raw_companies()

example_raw_weo()

example_raw_ipr()
}
\value{
A dataframe.
}
\description{
These datasets shows a state of the data prior to becoming an input for this
package.
}
\examples{
example_raw_companies()
example_raw_weo()
example_raw_ipr()
}
\keyword{internal}
