% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude.R
\name{exclude}
\alias{exclude}
\title{Exclude columns matching a pattern and the resulting duplicates}
\usage{
exclude(data, match)
}
\arguments{
\item{data}{A dataframe.}

\item{match}{A character vector. If length > 1, the union of the
matches is taken.

For \code{starts_with()}, \code{ends_with()}, and \code{contains()} this is an exact
match. For \code{matches()} this is a regular expression, and can be a
stringr pattern.}
}
\value{
A dataframe excluding the matching columns and duplicates.
}
\description{
Exclude columns matching a pattern and the resulting duplicates
}
\examples{
library(tibble)

# Excludes columns along with all its duplicates
data <- tibble(x = 1, y = 1:2)
data
data |> exclude("y")

# Columns are matched as a regular expression
data <- tibble(x = 1, yz = 1:2, zy = 1)
data
data |> exclude("y")
data |> exclude("y$")



# With a 'tilt_profile' excludes at both levels in a single step

product <- company <- tibble(companies_id = 1, y = "a", z = 1)
result <- tilt_profile(nest_levels(product, company))
result |> class()
result

out <- result |> exclude("y")
out |> unnest_product()
out |> unnest_company()
}
\concept{composable friends}
