% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extdata_path}
\alias{extdata_path}
\title{Get path to child extdata/}
\usage{
extdata_path(path)
}
\arguments{
\item{path}{Character. Path to a directory in inst/extdata/.}
}
\description{
Get path to child extdata/
}
\examples{
extdata_path("")
list.files(extdata_path(""), recursive = TRUE)
}
\keyword{internal}
