% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ictr_at_product_level.R
\name{ictr_at_product_level}
\alias{ictr_at_product_level}
\title{Aggregate rank and score for each input product}
\usage{
ictr_at_product_level(co2, low_threshold = 0.3, high_threshold = 0.7)
}
\arguments{
\item{co2}{A \link{data.frame} like \link{ictr_inputs}.}

\item{low_threshold}{A numeric value to segment low and medium transition
risk products.}

\item{high_threshold}{A numeric value to segment medium and high transition
risk products.}
}
\value{
A \link{data.frame}.
}
\description{
Input products will be ranked according to their footprint. Input products in
the highest percentile (≥70\%) will be classified as high transition risk
input products. Input products in the medium percentile (between ≥30\% and
<70\%) will be classified as medium transition risk input products. Products
in the lowest percentile (<30\%) will be classified as low transition risk
input products.

Therefore, each input product will receive either a score of 'high',
'medium', or 'low' based on the above mentioned classification criteria.
}
\examples{
ictr_at_product_level(ictr_inputs)
}
\seealso{
Other internal-ish functions: 
\code{\link{ictr_at_company_level}()},
\code{\link{pctr_at_company_level}()},
\code{\link{pctr_at_product_level}()}
}
\author{
Kalash Singhal.
}
\concept{internal-ish functions}
