% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ictr_data.R
\docType{data}
\name{ictr_companies}
\alias{ictr_companies}
\title{A dataset with fake companies, activity_uuid, and unit.}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 9 rows and 5 columns.
}
\source{
TODO
}
\usage{
ictr_companies
}
\description{
A dataset with fake companies, activity_uuid, and unit.
}
\examples{
ictr_companies
}
\seealso{
Other ICTR datasets: 
\code{\link{ictr_inputs}}
}
\author{
Kalash Singhal.
}
\concept{ICTR datasets}
\keyword{datasets}
