% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ictr_datasets}
\alias{ictr_datasets}
\alias{ictr_companies}
\alias{ictr_inputs}
\title{ICTR datasets}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9 rows and 3 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 33 rows and 5 columns.
}
\usage{
ictr_companies

ictr_inputs
}
\description{
ICTR datasets
}
\examples{
ictr_companies

ictr_inputs
}
\author{
Kalash Singhal.
}
\concept{ICTR datasets}
\keyword{datasets}
