% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ictr_score_companies.R
\name{ictr_score_companies}
\alias{ictr_score_companies}
\title{TODO}
\usage{
ictr_score_companies(scored, companies)
}
\arguments{
\item{scored}{TODO}

\item{companies}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
ictr_inputs |>
  ictr_score_inputs() |>
  ictr_score_companies(ictr_companies)
}
\seealso{
Other ICTR functions: 
\code{\link{ictr_score_inputs}()}
}
\concept{ICTR functions}
