% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ictr_score_inputs.R
\name{ictr_score_inputs}
\alias{ictr_score_inputs}
\title{TODO}
\usage{
ictr_score_inputs(inputs, low_threshold = 0.3, high_threshold = 0.7)
}
\arguments{
\item{inputs}{TODO}

\item{low_threshold}{TODO}

\item{high_threshold}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
library(dplyr)

# Minimum dataset
data <- ictr_toy_inputs()
data

ictr_score_inputs(data) |> glimpse()

# With fewer columns returns an error
data <- ictr_toy_inputs() |> select(-1)
try(ictr_score_inputs(data))

# Additional columns are added to the output
data <- ictr_toy_inputs() |> mutate(new = 1)
ictr_score_inputs(data) |> relocate(new)

# You may customize thresholds
data <- ictr_toy_inputs(input_co2 = c(1, 1.1, 2, 2.1, 3))
data

ictr_score_inputs(data, low_threshold = 1, high_threshold = 2)
}
\seealso{
Other ICTR functions: 
\code{\link{ictr_score_companies}()}
}
\concept{ICTR functions}
