% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ictr_data.R
\name{ictr_toy_inputs}
\alias{ictr_toy_inputs}
\title{Toy \code{ictr_inputs}}
\usage{
ictr_toy_inputs(input_co2 = 1, input_sector = "a", unit = "u", ...)
}
\arguments{
\item{input_co2, input_sector, unit}{See \link{ictr_inputs}.}

\item{...}{
  Arguments passed on to \code{\link[tibble:tibble]{tibble::tibble}}
  \describe{
    \item{\code{.rows}}{The number of rows, useful to create a 0-column tibble or
just as an additional check.}
    \item{\code{.name_repair}}{Treatment of problematic column names:
\itemize{
\item \code{"minimal"}: No name repair or checks, beyond basic existence,
\item \code{"unique"}: Make sure names are unique and not empty,
\item \code{"check_unique"}: (default value), no name repair, but check they are
\code{unique},
\item \code{"universal"}: Make the names \code{unique} and syntactic
\item a function: apply custom name repair (e.g., \code{.name_repair = make.names}
for names in the style of base R).
\item A purrr-style anonymous function, see \code{\link[rlang:as_function]{rlang::as_function()}}
}

This argument is passed on as \code{repair} to \code{\link[vctrs:vec_as_names]{vctrs::vec_as_names()}}.
See there for more details on these terms and the strategies used
to enforce them.}
  }}
}
\value{
A \link{data.frame}.
}
\description{
Toy \code{ictr_inputs}
}
\examples{
ictr_toy_inputs()
ictr_toy_inputs(unit = NULL)
ictr_toy_inputs(new = 1)
}
\seealso{
Other ICTR datasets: 
\code{\link{ictr_companies}},
\code{\link{ictr_inputs}}
}
\concept{ICTR datasets}
