% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inputs}
\alias{inputs}
\title{ICTR datasets}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 33 rows and 6 columns.
}
\usage{
inputs
}
\description{
ICTR datasets
}
\examples{
inputs
}
\author{
Kalash Singhal.
}
\concept{ICTR datasets}
\keyword{datasets}
