% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissions_profile.R
\name{insert_row_with_na_in_risk_category}
\alias{insert_row_with_na_in_risk_category}
\title{Insert row with NA in risk category}
\usage{
insert_row_with_na_in_risk_category(data)
}
\arguments{
\item{data}{A dataframe}
}
\value{
A dataframe
}
\description{
Insert row with NA in risk category
}
\keyword{internal}
