% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/istr.R, R/xctr_at_product_level.R
\name{istr}
\alias{istr}
\alias{istr_at_product_level}
\alias{istr_at_company_level}
\title{Calculate the ISTR indicator}
\usage{
istr(companies, scenario, mapper)

istr_at_product_level(companies, scenario, mapper)

istr_at_company_level(
  companies,
  co2,
  low_threshold = 1/3,
  high_threshold = 2/3
)
}
\arguments{
\item{companies}{A dataframe like \link{istr_companies}.}

\item{scenario}{A dataframe like \link{istr_weo_2022}.}

\item{mapper}{A dataframe like \link{istr_ep_weo}.}
}
\value{
At product level, a dataframe with at least columns companies_id, grouped_by, risk_category. At company level, a dataframe with at least columns companies_id, grouped_by, risk_category, value.
}
\description{
The Input Product Sector Risk Indicator assesses the transition risk of the
input products based on the sector's emissions targets the input product belongs
to. Those sector emission reduction targets vary across scenarios (e.g., net
zero scenario or 1.5° scenario) and the time horizon (e.g., reduction needed in
2030, 2040, 2050 to achieve the targets). It, therefore, is similar to the
Product Sector Risk Indicator but focuses on the input products that the company
needs to produce its products.The input products are, for example, resources,
packaging materials, energy and enabling services (such as tractor use on farm)
to produce the product.

If an input product belongs to a sector in which reduction targets are over 70\%,
it belongs to a high-risk sector and is classified as 'high sector transition
risk'. If the input product belongs to a sector in which reduction targets are
between 30-70\%, the product is classified into the category 'medium sector
transition risk' and if it is below 30\% it belongs to the category of 'low
sector transition risk'. If it does not belong to a sector that has
decarbonization targets, it is classified as 'no sector transition risk'.

After assessing the input products for each product, they are aggregated at
company-level to derive what percentage of the input products required by the
company to produce its products have high, medium and low sector transition
risk. We, therefore, derive the company-level information.

Please note that carbon emissions or emissions always mean CO2e.
}
\examples{
companies <- istr_companies
scenario <- istr_weo_2022
mapper <- istr_ep_weo

# Product level
companies |>
  istr_at_product_level(scenario, mapper)

# Company level
companies |>
  istr_at_product_level(scenario, mapper) |>
  istr_at_company_level()

# Same
istr(companies, scenario, mapper)
}
\concept{ISTR functions}
