% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/istr_add_reductions.R
\name{istr_add_reductions}
\alias{istr_add_reductions}
\title{Add the emission reduction targets to the companies dataset}
\usage{
istr_add_reductions(companies, weo_2022)
}
\arguments{
\item{companies}{A \link{data.frame}. The output of \code{\link[=istr_mapping]{istr_mapping()}}.}

\item{weo_2022}{A \link{data.frame} like \link{istr_weo_2022}.}
}
\value{
A dataframe with:
\itemize{
\item All the columns from the \code{companies} dataset.
\item New columns:
\itemize{
\item All the columns from the \code{ep_weo} dataset, but the columns
\code{eco_sectors} is named \code{ECO_sector} respectively.
\item All the columns from the \code{weo_2022} dataset, but the columns
\code{weo_product_mapper} and \code{weo_flow_mapper} are named \code{product} and
\code{flow}, respectively.
}
}
}
\description{
Adds the emission reduction values for each company's product(s).
}
\examples{
istr_companies |>
  istr_mapping(istr_ep_weo) |>
  istr_add_reductions(istr_weo_2022)
}
\seealso{
Other ISTR functions: 
\code{\link{istr_add_transition_risk}()},
\code{\link{istr_aggregate_scores}()},
\code{\link{istr_mapping}()}
}
\author{
Lyanne Ho.
}
\concept{ISTR functions}
