% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/istr_add_transition_risk.R
\name{istr_add_transition_risk}
\alias{istr_add_transition_risk}
\title{Categorize sector emission reduction targets}
\usage{
istr_add_transition_risk(with_reductions)
}
\arguments{
\item{with_reductions}{A data frame. The output of \code{\link[=istr_add_reductions]{istr_add_reductions()}}.}
}
\value{
A data frame with:
\itemize{
\item All the columns from the \code{companies} dataset.
\item All the columns from the \code{ep_weo} dataset, but the columns
\code{eco_sectors} is named \code{ECO_sector} respectively.
\item All the columns from the \code{weo_2022} dataset, but the columns
\code{weo_product_mapper} and \code{weo_flow_mapper} are named \code{product} and
\code{flow}, respectively.
\item New column:
\itemize{
\item \code{transition_risk} that holds the categorization of the \code{reductions}
column.
}
}
}
\description{
Translates the \code{reductions} column into three categories:
\itemize{
\item "low" if \code{reductions} <= 30.0
\item "medium" if 30.0 < \code{reductions} <= 70.0
\item "high" if \code{reductions} > 70.0
}
}
\examples{
istr_companies |>
  istr_add_reductions(istr_ep_weo, istr_weo_2022) |>
  istr_add_transition_risk()
}
\seealso{
Other ISTR functions: 
\code{\link{istr_add_reductions}()},
\code{\link{istr_aggregate_scores}()},
\code{\link{istr_mapping}()}
}
\concept{ISTR functions}
