% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/istr_aggregate_scores.R
\name{istr_aggregate_scores}
\alias{istr_aggregate_scores}
\title{Aggregate the products' scores for each company}
\usage{
istr_aggregate_scores(with_transition_risk, companies)
}
\arguments{
\item{with_transition_risk}{A \link{data.frame}. The output of
\code{\link[=istr_add_transition_risk]{istr_add_transition_risk()}}.}

\item{companies}{A \link{data.frame} like \link{istr_companies}.}
}
\value{
A \link{data.frame} with the columns:
\itemize{
\item \code{company_name}
\item \code{transition_risk}
\item \code{scenario}
\item \code{year}
\item \code{score_aggregated}, which holds the aggregated scores in percentage.
}
}
\description{
Calculates on a company-level the percentage of products that are in low /
medium / high transition risk.
}
\examples{
istr_companies |>
  istr_mapping(istr_ep_weo) |>
  istr_add_reductions(istr_weo_2022) |>
  istr_add_transition_risk() |>
  istr_aggregate_scores(istr_companies)
}
\seealso{
Other ISTR functions: 
\code{\link{istr_add_reductions}()},
\code{\link{istr_add_transition_risk}()},
\code{\link{istr_mapping}()}
}
\concept{ISTR functions}
