% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{istr_datasets}
\alias{istr_datasets}
\alias{istr_companies}
\alias{istr_ep_weo}
\alias{istr_weo_2022}
\title{ISTR datasets}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 33 rows and 36 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 133 rows and 6 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 8 rows and 10 columns.
}
\usage{
istr_companies

istr_ep_weo

istr_weo_2022
}
\description{
ISTR datasets
}
\examples{
istr_companies

istr_ep_weo

istr_weo_2022
}
\author{
Lyanne Ho.
}
\concept{ISTR datasets}
\keyword{datasets}
