% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/istr_data.R
\docType{data}
\name{istr_ep_weo}
\alias{istr_ep_weo}
\title{TODO}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 133 rows and 6 columns.
}
\usage{
istr_ep_weo
}
\description{
TODO
}
\examples{
istr_ep_weo
}
\seealso{
Other ISTR datasets: 
\code{\link{istr_companies}},
\code{\link{istr_weo_2022}}
}
\concept{ISTR datasets}
\keyword{datasets}
