% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/istr_mapping.R
\name{istr_mapping}
\alias{istr_mapping}
\title{Join the Ecoinvent's' sectors and the WEO 2022 file on the sectors by use the mapper that serves as a sector classification bridge between.}
\usage{
istr_mapping(companies, ep_weo)
}
\arguments{
\item{companies}{A \link{data.frame} like \link{istr_companies}.}

\item{ep_weo}{A \link{data.frame} like \link{istr_ep_weo}.}
}
\value{
A dataframe with:
\itemize{
\item All the columns from the \code{companies} dataset.
\item New columns:
\itemize{
\item All the columns from the \code{ep_weo} dataset, but the columns
\code{eco_sectors} is named \code{ECO_sector} respectively.
}
}
}
\description{
Join the Ecoinvent's' sectors and the WEO 2022 file on the sectors by use the mapper that serves as a sector classification bridge between.
}
\examples{
istr_companies |>
  istr_mapping(istr_ep_weo)
}
\seealso{
Other ISTR functions: 
\code{\link{istr_add_reductions}()},
\code{\link{istr_add_transition_risk}()},
\code{\link{istr_aggregate_scores}()}
}
\author{
Lyanne Ho.
}
\concept{ISTR functions}
