% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/istr_data.R
\docType{data}
\name{istr_weo_2022}
\alias{istr_weo_2022}
\title{TODO}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 8 rows and 10 columns.
}
\usage{
istr_weo_2022
}
\description{
TODO
}
\examples{
istr_weo_2022
}
\seealso{
Other ISTR datasets: 
\code{\link{istr_companies}},
\code{\link{istr_ep_weo}}
}
\author{
Lyanne Ho.
}
\concept{ISTR datasets}
\keyword{datasets}
