% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jitter_range.R
\name{jitter_range}
\alias{jitter_range}
\title{Expand a range adding some random noise}
\usage{
jitter_range(data, ...)
}
\arguments{
\item{data}{A dataframe with columns \code{min} and \code{max}.}

\item{...}{Arguments passed to \code{\link[=jitter]{jitter()}}.}
}
\value{
The input dataframe with the additional columns \code{min_jitter} and
\code{max_jitter}.
}
\description{
This function expands a range by adding noise to the left of the minimum
values and to the right of the maximum values.
}
\examples{
library(tibble)

data <- tibble(min = 1:2, max = 3:4)

data |> jitter_range()

data |> jitter_range(amount = 0.9)
}
\seealso{
Other helpers: 
\code{\link{summarize_range}()},
\code{\link{unnest_product}()}
}
\concept{helpers}
