% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_to.R
\name{join_to}
\alias{join_to}
\title{Join every possible column and row}
\usage{
join_to(x, y)
}
\arguments{
\item{x, y}{A pair of data frames, data frame extensions (e.g. a tibble), or
lazy data frames (e.g. from dbplyr or dtplyr). See \emph{Methods}, below, for
more details.}
}
\value{
A data frame with all columns in \code{x} and \code{y} and all rows in \code{y},
except the columns and resulting duplicates matched by the argument
\code{excluding}.
}
\description{
This function is typically useful when you want to pipe data into a summary
and then join back the data to the summary. It joins by all shared columns
using a "many-to-many" relationship.
}
\examples{
library(dplyr, warn.conflicts = FALSE)

data <- tibble(x = rep(1, 4), y = letters[rep(c(1, 2), 2)], z = 1:4)
data

data |>
  summarize(mean = mean(x), .by = "y") |>
  join_to(data)

data |>
  summarize(mean = mean(x), .by = "y") |>
  join_to(data |> exclude("z"))
}
\seealso{
Other helpers: 
\code{\link{jitter_range}()},
\code{\link{summarize_range}()},
\code{\link{unnest_product}()}
}
\concept{helpers}
