% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvp_path.R
\name{mvp_path}
\alias{mvp_path}
\alias{mvp_paths}
\title{Get path to MVPs}
\usage{
mvp_path(path)

mvp_paths(pattern = NULL)
}
\arguments{
\item{path}{Name of file.}

\item{pattern}{A regular expression of filenames to match. If NULL all
available files are returned. listed.}
}
\value{
character
}
\description{
The minimal-viable-product (MVP) of each indicator was originally developed
in an .Rmd file that we gradually refactored towards production. They are
stored in inst/extdata/mvp/ in their original form -- except for minimal
changes such as redirection of inputs and outputs to avoid leaking private
data.
}
\examples{
mvp_paths()
mvp_path("pstr.Rmd")
try(mvp_path())
try(mvp_path("inexistent.file"))
}
\keyword{internal}
