% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvp_sample_dataset_demo.R
\docType{data}
\name{mvp_sample_dataset_demo}
\alias{mvp_sample_dataset_demo}
\title{A public dataset}
\format{
An object of class \code{fs_path} (inherits from \code{character}) of length 1.
}
\source{
Tilman via https://github.com/2DegreesInvesting/tiltData/issues/200
}
\usage{
mvp_sample_dataset_demo
}
\description{
Public based on a meeting on 2023-03-22
}
\examples{
mvp_sample_dataset_demo
}
\keyword{datasets}
