% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_levels.R
\name{nest_levels}
\alias{nest_levels}
\title{Nest levels}
\usage{
nest_levels(product, company)
}
\description{
Nest levels
}
\examples{
product <- tibble::tibble(companies_id = 1, x = 1)
company <- tibble::tibble(companies_id = 1, x = 1)
nest_levels(product, company)
}
\keyword{internal}
