% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pctr_at_company_level.R
\name{pctr_at_company_level}
\alias{pctr_at_company_level}
\title{Aggregate the products' scores for each company}
\usage{
pctr_at_company_level(scored_activities, companies)
}
\arguments{
\item{scored_activities}{A \link{data.frame}. The output of
\link{pctr_at_product_level}.}

\item{companies}{A \link{data.frame} like \link{pctr_companies}.}
}
\value{
A \link{data.frame}.
}
\description{
Calculates on a company-level the percentage of products that are in low /
medium / high transition risk due to the products' relative carbon footprint.

Activities are mapped with the companies' products using common columns
present in dataframes \code{scored_activities} and \code{companies}
}
\examples{
library(dplyr, warn.conflicts = FALSE)

one_company <- pctr_companies |> filter(company_id \%in\% first(company_id))
one_company

pctr_ecoinvent_co2 |>
  pctr_at_product_level(low_threshold = 0.3, high_threshold = 0.7) |>
  pctr_at_company_level(one_company)
}
\seealso{
Other internal-ish functions: 
\code{\link{ictr_at_company_level}()},
\code{\link{ictr_at_product_level}()},
\code{\link{pctr_at_product_level}()}
}
\author{
Tilman Trompke.
}
\concept{internal-ish functions}
