% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xctr.R
\name{pctr_at_product_level}
\alias{pctr_at_product_level}
\title{Deprecated}
\usage{
pctr_at_product_level(
  companies,
  co2,
  low_threshold = 1/3,
  high_threshold = 2/3
)
}
\description{
Deprecated
}
\keyword{internal}
