% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pctr_data.R
\docType{data}
\name{pctr_companies}
\alias{pctr_companies}
\title{TODO add title}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 25 rows and 6 columns.
}
\source{
TODO add source
}
\usage{
pctr_companies
}
\description{
TODO add title
}
\examples{
pctr_companies
}
\seealso{
Other datasets for product carbon transition risk: 
\code{\link{pctr_ecoinvent_co2}}
}
\concept{datasets for product carbon transition risk}
\keyword{datasets}
