% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pctr_company_scores.R
\name{pctr_company_scores}
\alias{pctr_company_scores}
\title{Typically one sentence like "Given an INPUT returns an OUTPUT"}
\usage{
pctr_company_scores(path)
}
\arguments{
\item{data}{(Typically one or more data frames with crucial columns)}
}
\value{
(Typically one or more data frames with crucial columns)
}
\description{
A short introduction, typically one paragraph but maybe more.
}
\details{
A longer explanation, typically multiple paragraphs but maybe one.
}
\examples{
# TODO
}
