% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pctr_datasets}
\alias{pctr_datasets}
\alias{pctr_companies}
\alias{pctr_ecoinvent_co2}
\title{PCTR datasets}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9 rows and 4 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7 rows and 6 columns.
}
\usage{
pctr_companies

pctr_ecoinvent_co2
}
\description{
PCTR datasets
}
\examples{
pctr_companies

pctr_ecoinvent_co2
}
\author{
Tilman Trompke.
}
\concept{PCTR datasets}
\keyword{datasets}
