% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pctr_data.R
\docType{data}
\name{pctr_ecoinvent_co2}
\alias{pctr_ecoinvent_co2}
\title{TODO add title}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 25 rows and 9 columns.
}
\source{
TODO add source
}
\usage{
pctr_ecoinvent_co2
}
\description{
TODO add title
}
\examples{
pctr_ecoinvent_co2
}
\seealso{
Other PCTR datasets: 
\code{\link{pctr_companies}}
}
\concept{PCTR datasets}
\keyword{datasets}
