% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pctr_company_scores.R
\name{pctr_impl}
\alias{pctr_impl}
\title{\subsection{Introduction}{
}}
\usage{
pctr_impl(data_ecoinvent, data_companies)
}
\description{
The Product Carbon Transition Risk measures the transition risk on the
product-level of a company. The indicator is expressed in the percentage
of all products that are at high risk, medium risk or low risk due to
the products’ relative carbon footprint. The assessment is first
performed on a product-level and can then be aggregated to the
company-level.
}
\details{
The Product Carbon Transition Risk measures the relative carbon
footprint per product. As a default option each product will be compared
to the carbon footprint of every other product. Alternatively, users can
also choose to compare those products of the same type, so e.g.,
comparing only raw materials with each other. Products with a higher
carbon-footprint will also face a higher risk.

After identifying each carbon footprint for one product, the products
will be ranked according to their carbon footprint. Products in the
highest percentile (≥70\%) will be classified as high transition risk
products. Products in the medium percentile (between ≥30\% and \<70\%)
will be classified as medium transition risk products. Products in the
lowest percentile (\<30\%) will be classified as low transition risk
products.

After categorisation, we will aggregate all products from the same
category and set them in relation to all products that the company
produces. We derive the Product Carbon Transition Risks.

Please note that carbon footprints, and emissions are used equivalently.
Carbon footprint refers only on the emissions which occur at the
production stage of the product and not the emissions from the inputs.
The unit is CO2e in kg.

The goal of the transition risk MVP is to create a first draft of how
the transition risk indicator for products would be build in code in
order to convert it into code production easily in the future. This MVP
should provide the share of products with “low”, “medium”, or “high”
relative production emissions per company.

In this markdown file the aim is to prepare data so that they contain
the following:
\itemize{
\item a column with production emissions
\item a column indicating the percentile relative to (i) all products with
same unit (ii) products in same sector (iii) products in same segment
\item a column indicating whether the product has “low”, “medium” or “high”
relative production emissions
}

The product carbon transition risk indicator consists of 2 main steps:

\if{html}{\out{<div class="sourceCode">}}\preformatted{- Step 1: Calculating the relative carbon footprint per product
- Step 2: Aggregating on the company-level
}\if{html}{\out{</div>}}
\subsection{Step 1: Calculating the relative carbon footprint per product}{
}

\subsection{Data}{
}

\subsection{Prepare data so that they contain the respective columns:}{
\itemize{
\item replace sample activities for companies with “real activities” from
Hanish’s sample
\item include absolute score
\item use ecoinvent’s v3.9 data
}
}

\subsection{1. Load packages and ecoinvent data}{
}

\subsection{2. Create subset of data with relevant columns}{
}

\subsection{3. Add column “production_emissions_per_product”}{
}

\subsection{B) Calculation of transition risks}{
}

\subsection{1. Calculate percentiles in comparison to different sub-sets}{
}

\subsection{2. Assign scores to the activities based on their position within the distribution}{
\subsection{Step 2: Aggregating on the company-level}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{  Aggregating on the company-level the percentage of transistion risk products for each company.
  Calculate 'Product Carbon TR Indicator' for example companies from Hanish's compiled dataset
}\if{html}{\out{</div>}}
}

}

\subsection{Data}{
}

\subsection{1. Loading sample company data set & create sample product list}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{NOTE:

- Problem A with dataset: in the MVP_company_sample dataset the activity names from ecoinvent have an underscore "\\_" instead
  of space in their name. E.g., there is an activity called "treatment_of_tyre_wear_emissions_passenger_car" instead of            "treatment of tyre wear emissions, passenger car" as in the original dataset.

- Quick solution for now (I don't want to invest time in replacing underscores with spaces and vice versa: I replace the            products with random example from ecoinvent

- Problem B with dataset: in the sample MVP dataset, some company_name were associated with multiple company_ids (example:          instead of ID number 00001 for each row, excel filled in 00002, 00003, 00004 in the following rows). I adjusted that            manually, i.e. the issue is resolved in the following dataset.
}\if{html}{\out{</div>}}
}

\subsection{2. Match data_short and data_companies so that data_companies includes emissions columns + scores}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{Result: we end up with a dataset 'companies_emissions' consisting of 1257 observations. This is because the ecoinvent            database consists of multiple rows with the same activity_name & unit.
}\if{html}{\out{</div>}}
}

\subsection{3. Caclulate the scores for each company (i.e. shares for each company)}{

\#Question: How to include results for different scores in one dataset,
i.e. not only grouping by score_all but also all other scores?
}

\subsection{4. Create final dataset: should have three rows per company, i.e. one for each risk_group low-RCF, medium-RCF, high-RCF. Column risk_shares indicates per risk_group the \% of products of the respective company.}{
}
}
