% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pctr.R
\name{pctr_path}
\alias{pctr_path}
\title{Path to indicator}
\usage{
pctr_path(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[fs:path]{fs::path}}
  \describe{
    \item{\code{ext}}{An optional extension to append to the generated path.}
  }}
}
\value{
A path
}
\description{
Path to indicator
}
\examples{
pctr_path()
fs::dir_tree(pctr_path())
}
\keyword{internal}
