% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pctr_score_activities.R
\name{pctr_score_activities}
\alias{pctr_score_activities}
\title{Title TODO}
\usage{
pctr_score_activities(ecoinvent_co2, low_threshold = 0.3, high_threshold = 0.7)
}
\arguments{
\item{ecoinvent_co2}{A \link{data.frame} like \link{pctr_ecoinvent_co2}.}

\item{low_threshold}{A numeric value of length 1 giving ... TODO.}

\item{high_threshold}{A numeric value of length 1 giving ... TODO.}
}
\value{
A \link{data.frame} with columns ... TODO.
}
\description{
Title TODO
}
\examples{
pctr_ecoinvent_co2 |>
  pctr_score_activities()
}
\seealso{
Other PCTR functions: 
\code{\link{pctr_score_companies}()}
}
\concept{PCTR functions}
