% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jitter_range.R
\name{percent_noise}
\alias{percent_noise}
\title{Calculate percent noise}
\usage{
percent_noise(x, noisy)
}
\arguments{
\item{x}{Numeric vector.}

\item{noisy}{Numeric vector.}
}
\value{
Numeric vector.
}
\description{
Calculate percent noise
}
\examples{
x <- -10:10
noisy <- jitter(x)
out <- percent_noise(x, noisy)
out

finite <- out[is.finite(out)]
mean(finite)
}
\keyword{internal}
