% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{products}
\alias{products}
\title{PCTR datasets}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 6 columns.
}
\usage{
products
}
\description{
PCTR datasets
}
\examples{
products
}
\author{
Tilman Trompke.
}
\concept{PCTR datasets}
\keyword{datasets}
