% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstr_at_product_level.R, R/sector_profile.R,
%   R/xctr.R
\name{pstr_at_product_level}
\alias{pstr_at_product_level}
\alias{pstr}
\alias{pstr_at_company_level}
\title{Calculate the sector profile indicator}
\usage{
pstr_at_product_level(
  companies,
  scenarios,
  low_threshold = ifelse(scenarios$year == 2030, 1/9, 1/3),
  high_threshold = ifelse(scenarios$year == 2030, 2/9, 2/3)
)

pstr(
  companies,
  scenarios,
  low_threshold = ifelse(scenarios$year == 2030, 1/9, 1/3),
  high_threshold = ifelse(scenarios$year == 2030, 2/9, 2/3)
)

pstr_at_company_level(data)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{pstr()} was renamed to \code{sector_profile()}.
}
\keyword{internal}
