% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstr_add_reductions.R
\name{pstr_add_reductions}
\alias{pstr_add_reductions}
\title{Title}
\usage{
pstr_add_reductions(companies, ep_weo, weo_2022)
}
\arguments{
\item{companies}{TODO}

\item{ep_weo}{TODO}

\item{weo_2022}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{
# TODO
}
