% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstr_add_reductions.R,
%   R/pstr_add_reductions_old.R
\name{pstr_add_reductions_old}
\alias{pstr_add_reductions_old}
\title{Add the emission reduction targets to the companies dataset}
\usage{
pstr_add_reductions_old(companies, ep_weo, weo_2022)

pstr_add_reductions_old(companies, ep_weo, weo_2022)
}
\arguments{
\item{companies}{A \link{data.frame} like \link{pstr_companies}.}

\item{ep_weo}{A \link{data.frame} like \link{pstr_ep_weo}.}

\item{weo_2022}{A \link{data.frame} like \link{pstr_weo_2022}.}
}
\value{
A dataframe with:
\itemize{
\item All the columns from the \code{companies} dataset.
\item New columns:
\itemize{
\item All the columns from the \code{ep_weo} dataset, but the columns
\code{EP_sector} and \code{EP_subsector} are named \code{sector} and \code{subsector}
respectively.
}
}

A dataframe with:
\itemize{
\item All the columns from the \code{companies} dataset.
\item New columns:
\itemize{
\item All the columns from the \code{ep_weo} dataset, but the columns
\code{EP_sector} and \code{EP_subsector} are named \code{sector} and \code{subsector}
respectively.
}
}
}
\description{
Adds the emission reduction values for each company's product(s).

Adds the emission reduction values for each company's product(s).
}
\examples{
pstr_companies |>
  pstr_add_reductions_old(pstr_ep_weo, pstr_weo_2022)
pstr_companies |>
  pstr_add_reductions_old(pstr_ep_weo, pstr_weo_2022)
}
\seealso{
Other PSTR functions: 
\code{\link{pstr_add_transition_risk}()},
\code{\link{pstr_aggregate_scores}()}

Other PSTR functions: 
\code{\link{pstr_add_transition_risk}()},
\code{\link{pstr_aggregate_scores}()}
}
\author{
Linda Delacombaz.
}
\concept{PSTR functions}
