% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstr_add_transition_risk.R
\name{pstr_add_transition_risk}
\alias{pstr_add_transition_risk}
\title{Title}
\usage{
pstr_add_transition_risk(with_reductions)
}
\arguments{
\item{with_reductions}{A data frame. The output of \code{\link[=pstr_add_reductions]{pstr_add_reductions()}}.}
}
\value{
TODO
}
\description{
Title
}
\examples{
# TODO
}
