% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstr_aggregate_scores.R
\name{pstr_aggregate_scores}
\alias{pstr_aggregate_scores}
\title{Title}
\usage{
pstr_aggregate_scores(with_transition_risk, companies)
}
\arguments{
\item{with_transition_risk}{A data frame. The output of
\code{\link[=pstr_add_transition_risk]{pstr_add_transition_risk()}}.}

\item{companies}{A \code{companies} dataframe like \code{\link{companies}}.}
}
\value{
TODO
}
\description{
Title
}
\examples{
# TODO
}
