% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstr_at_company_level.R
\name{pstr_at_company_level}
\alias{pstr_at_company_level}
\title{Aggregate the products' scores for each company}
\usage{
pstr_at_company_level(with_transition_risk, companies)
}
\arguments{
\item{with_transition_risk}{A \link{data.frame}. The output of
\code{\link[=pstr_at_product_level]{pstr_at_product_level()}}.}
}
\value{
A \link{data.frame} with the columns:
\itemize{
\item \code{company_name}
\item \code{transition_risk}
\item \code{scenario}
\item \code{year}
\item \code{score_aggregated}, which holds the aggregated scores in percentage.
}
}
\description{
Calculates on a company-level the percentage of products that are in low /
medium / high transition risk.
}
\examples{
companies <- pstr_companies[1, ]
scenarios <- pstr_scenarios

companies |>
  pstr_at_product_level(scenarios) |>
  pstr_at_company_level(companies)
}
\seealso{
Other PSTR functions: 
\code{\link{pstr}()}
}
\author{
Linda Delacombaz.
}
\concept{PSTR functions}
\keyword{internal}
