% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstr_at_product_level.R
\name{pstr_at_product_level}
\alias{pstr_at_product_level}
\title{Calculate the PCTR at product level}
\usage{
pstr_at_product_level(companies, scenarios)
}
\arguments{
\item{companies}{A dataframe like \link{pstr_companies}.}

\item{scenarios}{A dataframe like \link{pstr_scenarios}.}
}
\value{
A dataframe.
}
\description{
Calculate the PCTR at product level
}
\examples{
pstr_at_product_level(pstr_companies, pstr_scenarios)
}
\keyword{internal}
