% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstr_data.R
\docType{data}
\name{pstr_companies}
\alias{pstr_companies}
\title{A dataset with fake companies, products, sectors and subsectors}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 28 rows and 9 columns.
}
\source{
Companies, sectors, subsectors and products from Europages.
}
\usage{
pstr_companies
}
\description{
A dataset with fake companies, products, sectors and subsectors
}
\examples{
pstr_companies
}
\seealso{
Other PSTR datasets: 
\code{\link{pstr_scenarios}}
}
\author{
Tilman Trompke.
}
\concept{PSTR datasets}
\keyword{datasets}
