% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pstr_datasets}
\alias{pstr_datasets}
\alias{pstr_companies}
\alias{pstr_scenarios}
\title{PSTR datasets}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 28 rows and 10 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 484 rows and 10 columns.
}
\source{
\itemize{
\item Companies, sectors, subsectors and products from Europages.
\item WEO data from the \href{https://www.iea.org/}{International Energy Agency} (IEA)'s
\href{https://www.iea.org/topics/world-energy-outlook}{World Energy Outlook} (WEO)
2022 scenario on the global level.
}
}
\usage{
pstr_companies

pstr_scenarios
}
\description{
PSTR datasets
}
\examples{
pstr_companies

pstr_scenarios
}
\author{
Tilman Trompke.
}
\concept{PSTR datasets}
\keyword{datasets}
