% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstr_data.R
\docType{data}
\name{pstr_ep_weo}
\alias{pstr_ep_weo}
\title{A sector mapper between Europages and \code{weo_2022}.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 133 rows and 6 columns.
}
\source{
Internal source.
}
\usage{
pstr_ep_weo
}
\description{
A sector mapper between Europages and \code{weo_2022}.
}
\examples{
pstr_ep_weo
}
\seealso{
Other PSTR datasets: 
\code{\link{pstr_companies}},
\code{\link{pstr_new_companies}},
\code{\link{pstr_new_ipr_2022}},
\code{\link{pstr_new_weo_2022}},
\code{\link{pstr_weo_2022}}
}
\author{
Linda Delacombaz.
}
\concept{PSTR datasets}
\keyword{datasets}
