% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstr_new_add_reductions.R
\name{pstr_new_add_reductions}
\alias{pstr_new_add_reductions}
\title{Add the emission reduction targets to the companies dataset}
\usage{
pstr_new_add_reductions(companies, scenarios)
}
\arguments{
\item{companies}{TODO}

\item{scenario}{TODO}
}
\value{
A dataframe with:
\itemize{
\item All the columns from the \code{companies} dataset.
\item New columns:
\itemize{
\item All the columns from the \code{scenario} dataset.
}
}
}
\description{
Adds the emission reduction values for each company's product(s).
}
\examples{
library(dplyr, warn.conflicts = FALSE)

weo <- slice(pstr_new_weo_2022, 1)
ipr <- slice(pstr_new_ipr_2022, 1)
scenarios <- pstr_prepare_scenario(list(weo = weo, ipr = ipr))

companies <- pstr_prepare_companies(slice(pstr_new_companies, 1))

pstr_new_add_reductions(companies, scenarios)
}
\seealso{
Other PSTR functions: 
\code{\link{pstr_add_reductions}()},
\code{\link{pstr_add_transition_risk}()},
\code{\link{pstr_aggregate_scores}()},
\code{\link{pstr}()}
}
\concept{PSTR functions}
