% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstr_new_data.R
\docType{data}
\name{pstr_new_ipr_2022}
\alias{pstr_new_ipr_2022}
\title{A dataset with scenario, year, product and reductions information}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 112 rows and 7 columns.
}
\source{
TODO
}
\usage{
pstr_new_ipr_2022
}
\description{
\itemize{
\item TODO: Where does it come from and what wrangling did you do?
}
}
\examples{
pstr_new_ipr_2022
}
\seealso{
Other PSTR datasets: 
\code{\link{pstr_companies}},
\code{\link{pstr_ep_weo}},
\code{\link{pstr_new_companies}},
\code{\link{pstr_new_weo_2022}},
\code{\link{pstr_weo_2022}}
}
\author{
Tilman Trompke.
}
\concept{PSTR datasets}
\keyword{datasets}
