% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstr_add_reductions.R
\name{pstr_old_add_reductions}
\alias{pstr_old_add_reductions}
\title{Add the emission reduction targets to the companies dataset}
\usage{
pstr_old_add_reductions(companies, ep_weo, weo_2022)
}
\arguments{
\item{companies}{A \link{data.frame} like \link{pstr_old_companies}.}

\item{ep_weo}{A \link{data.frame} like \link{pstr_ep_weo}.}

\item{weo_2022}{A \link{data.frame} like \link{pstr_weo_2022}.}
}
\value{
A dataframe with:
\itemize{
\item All the columns from the \code{companies} dataset.
\item New columns:
\itemize{
\item All the columns from the \code{ep_weo} dataset, but the columns
\code{EP_sector} and \code{EP_subsector} are named \code{sector} and \code{subsector}
respectively.
}
}
}
\description{
Adds the emission reduction values for each company's product(s).
}
\examples{
pstr_old_companies |>
  pstr_old_add_reductions(pstr_ep_weo, pstr_weo_2022)
}
\seealso{
Other PSTR functions: 
\code{\link{pstr_add_transition_risk}()},
\code{\link{pstr_at_company_level}()},
\code{\link{pstr}()}
}
\author{
Linda Delacombaz.
}
\concept{PSTR functions}
\keyword{internal}
