% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xstr_polish_output_at_copmany_level.R
\name{pstr_polish_output_at_copmany_level}
\alias{pstr_polish_output_at_copmany_level}
\title{Polish output at company level}
\usage{
pstr_polish_output_at_copmany_level(data)
}
\arguments{
\item{data}{The output of \verb{xstr*()} functions.}
}
\value{
A dataframe.
}
\description{
Polish output at company level
}
\examples{
library(dplyr, warn.conflicts = FALSE)

out <- pstr(slice(pstr_companies, 1), pstr_scenarios)
out

out |> pstr_polish_output_at_copmany_level()
}
\seealso{
Other helpers: 
\code{\link{pstr_prepare_scenario}()},
\code{\link{xstr_pivot_type_sector_subsector}()}
}
\concept{helpers}
