% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstr_prepare_companies.R
\name{pstr_prepare_companies}
\alias{pstr_prepare_companies}
\title{Given a messy PSTR \code{companies} dataframe returns a cleaner one}
\usage{
pstr_prepare_companies(data)
}
\arguments{
\item{data}{A dataframe with these columns:
\itemize{
\item ipr_sector
\item ipr_subsector
\item weo_product
\item weo_flow
}}
}
\value{
A \code{copmanies} dataset as required by PSTR functions.
}
\description{
Given a messy PSTR \code{companies} dataframe returns a cleaner one
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(readr, warn.conflicts = FALSE)

raw_companies <- read_csv(extdata_path("pstr_companies.csv"))
glimpse(raw_companies)

companies <- pstr_prepare_companies(raw_companies)
companies
}
\seealso{
Other pre-processing helpers: 
\code{\link{pstr_prepare_scenario}()}
}
\concept{pre-processing helpers}
