% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstr_prepare_scenario.R
\name{pstr_prepare_scenario}
\alias{pstr_prepare_scenario}
\title{Prepare the scenario data set}
\usage{
pstr_prepare_scenario(ipr, weo)
}
\arguments{
\item{ipr}{TODO}

\item{weo}{TODO}
}
\value{
A dataframe with:
\itemize{
\item All the columns from the \code{ipr} dataset.
\item All the columns from the \code{weo} dataset.
}
}
\description{
Prepare the scenario data set
}
\examples{
#TODO
}
\seealso{
Other PSTR functions: 
\code{\link{pstr_add_reductions}()},
\code{\link{pstr_add_transition_risk}()},
\code{\link{pstr_aggregate_scores}()},
\code{\link{pstr_new_add_reductions}()},
\code{\link{pstr_prepare_companies}()},
\code{\link{pstr}()}
}
\concept{PSTR functions}
\keyword{internal}
