% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstr_raw_data.R
\docType{data}
\name{pstr_raw_companies}
\alias{pstr_raw_companies}
\title{A dataset with fake companies, products, sectors and subsectors}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 17 rows and 14 columns.
}
\source{
Companies, sectors, subsectors and products from Europages.
}
\usage{
pstr_raw_companies
}
\description{
A dataset with fake companies, products, sectors and subsectors
}
\examples{
pstr_raw_companies
}
\seealso{
Other PSTR datasets: 
\code{\link{pstr_old_companies}},
\code{\link{pstr_raw_ipr_2022}},
\code{\link{pstr_raw_weo_2022}}
}
\author{
Tilman Trompke.
}
\concept{PSTR datasets}
\keyword{datasets}
