% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstr_raw_data.R
\docType{data}
\name{pstr_raw_weo_2022}
\alias{pstr_raw_weo_2022}
\title{A dataset with scenario, year, product and reductions information}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 372 rows and 8 columns.
}
\source{
Data from the \href{https://www.iea.org/}{International Energy Agency}
(IEA)'s \href{https://www.iea.org/topics/world-energy-outlook}{World Energy Outlook}
(WEO) 2022 scenario on the global level.
}
\usage{
pstr_raw_weo_2022
}
\description{
\itemize{
\item TODO: Where does it come from and what wrangling did you do?
}
}
\examples{
pstr_raw_weo_2022
}
\seealso{
Other PSTR datasets: 
\code{\link{pstr_companies}},
\code{\link{pstr_ep_weo}},
\code{\link{pstr_raw_companies}},
\code{\link{pstr_raw_ipr_2022}},
\code{\link{pstr_weo_2022}}
}
\author{
Tilman Trompke.
}
\concept{PSTR datasets}
\keyword{datasets}
