% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstr_data.R
\docType{data}
\name{pstr_scenarios}
\alias{pstr_scenarios}
\title{A dataset with data from multiple scenarios}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 484 rows and 11 columns.
}
\source{
WEO data from the \href{https://www.iea.org/}{International Energy
Agency} (IEA)'s
\href{https://www.iea.org/topics/world-energy-outlook}{World Energy
Outlook} (WEO) 2022 scenario on the global level.
}
\usage{
pstr_scenarios
}
\description{
A dataset with data from multiple scenarios
}
\examples{
pstr_scenarios
}
\seealso{
Other PSTR datasets: 
\code{\link{pstr_companies}}
}
\author{
Tilman Trompke.
}
\concept{PSTR datasets}
\keyword{datasets}
