% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstr_data.R
\docType{data}
\name{pstr_weo_2022}
\alias{pstr_weo_2022}
\title{A dataset with scenario, year, product and reductions information}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8 rows and 10 columns.
}
\source{
Data from IEA's WEO scenarios.
}
\usage{
pstr_weo_2022
}
\description{
\itemize{
\item TODO: Where does it come from and what wrangling did you do?
}
}
\examples{
pstr_weo_2022
}
\seealso{
Other PSTR datasets: 
\code{\link{pstr_companies}},
\code{\link{pstr_ep_weo}}
}
\author{
Linda Delacombaz.
}
\concept{PSTR datasets}
\keyword{datasets}
