% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sector_profile.R
\name{sector_profile}
\alias{sector_profile}
\title{Calculate the indicator "sector profile"}
\usage{
sector_profile(
  companies,
  scenarios,
  low_threshold = ifelse(scenarios$year == 2030, 1/9, 1/3),
  high_threshold = ifelse(scenarios$year == 2030, 2/9, 2/3)
)
}
\arguments{
\item{companies}{A dataframe like \link{pstr_companies}.}

\item{scenarios}{A dataframe like \link{xstr_scenarios}.}

\item{low_threshold}{A numeric value to segment low and medium reduction
targets.}

\item{high_threshold}{A numeric value to segment medium and high reduction
targets.}
}
\value{
A data frame with the column \code{companies_id}, and the nested columns \code{product} and \code{company} holding the outputs at product and company level. Unnesting \code{product} yields a data frame with at least columns \code{companies_id}, \code{grouped_by}, \code{risk_category}. Unnesting \code{company} yields a data frame with at least columns \code{companies_id}, \code{grouped_by}, \code{risk_category}, \code{value}.
}
\description{
\if{html}{\out{<div class="sourceCode {r child=extdata_path("child/intro-sector-profile.Rmd")}">}}\preformatted{}\if{html}{\out{</div>}}
}
\examples{
library(tiltToyData)
library(readr)
options(readr.show_col_types = FALSE)

toy_files()

companies <- read_csv(toy_path("sector_profile_companies.csv.gz"))
scenarios <- read_csv(toy_path("sector_profile_any_scenarios.csv.gz"))

both <- sector_profile(companies, scenarios)
both

both |> unnest_product()

both |> unnest_company()
}
\concept{sector profile functions}
