% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/sector_profile_any_polish_output_at_copmany_level.R
\name{sector_profile_any_polish_output_at_company_level}
\alias{sector_profile_any_polish_output_at_company_level}
\title{Polish output at company level}
\usage{
sector_profile_any_polish_output_at_company_level(data)
}
\arguments{
\item{data}{The output of \verb{sector_profile*()} functions.}
}
\value{
A dataframe.
}
\description{
Polish output at company level
}
\examples{
library(tiltToyData)
library(readr)
options(readr.show_col_types = FALSE)

companies <- read_csv(toy_sector_profile_companies())
scenarios <- read_csv(toy_sector_profile_any_scenarios())

sector_profile(companies, scenarios) |>
  unnest_company() |>
  sector_profile_any_polish_output_at_company_level()

companies_upstream <- read_csv(toy_sector_profile_upstream_companies())
inputs <- read_csv(toy_sector_profile_upstream_products())

sector_profile_upstream(companies_upstream, scenarios, inputs) |>
  unnest_company() |>
  sector_profile_any_polish_output_at_company_level()
}
\concept{post-processing helpers}
