% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sector_profile_any_prepare_scenario.R
\name{sector_profile_any_prepare_scenario}
\alias{sector_profile_any_prepare_scenario}
\title{Given a named list of scenarios returns a cleaner \code{scenarios} dataframe}
\usage{
sector_profile_any_prepare_scenario(scenarios)
}
\arguments{
\item{scenarios}{A named list of identically structured scenarios.}
}
\value{
A single, cleaner dataframe with an additional column to identify
which rows come from which scenario.
}
\description{
Given a named list of scenarios returns a cleaner \code{scenarios} dataframe
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(readr, warn.conflicts = FALSE)

raw_weo <- read_csv(extdata_path("str_weo_targets.csv"))
raw_ipr <- read_csv(extdata_path("str_ipr_targets.csv"))
raw_scenarios <- list(weo = raw_weo, ipr = raw_ipr)

sector_profile_any_prepare_scenario(raw_scenarios)
}
\seealso{
Other pre-processing helpers: 
\code{\link{sector_profile_any_pivot_type_sector_subsector}()},
\code{\link{sector_profile_any_prune_companies}()}
}
\concept{pre-processing helpers}
