% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sector_profile_upstream.R
\name{sector_profile_upstream}
\alias{sector_profile_upstream}
\title{Calculate the indicator "sector profile upstream"}
\usage{
sector_profile_upstream(
  companies,
  scenarios,
  inputs,
  low_threshold = ifelse(scenarios$year == 2030, 1/9, 1/3),
  high_threshold = ifelse(scenarios$year == 2030, 2/9, 2/3)
)
}
\arguments{
\item{companies, scenarios, inputs}{A dataframe like the dataset with a matching name in tiltToyData (see \href{https://2degreesinvesting.github.io/tiltToyData/reference/index.html}{Reference}).}

\item{low_threshold}{A numeric value to segment low and medium transition
risk products.}

\item{high_threshold}{A numeric value to segment medium and high transition
risk products.}
}
\value{
A data frame with the column \code{companies_id}, and the nested columns\code{product} and \code{company} holding the outputs at product and company level. Unnesting \code{product} yields a data frame with at least columns \code{companies_id}, \code{grouped_by}, \code{risk_category}. Unnesting \code{company} yields a data frame with at least columns \code{companies_id}, \code{grouped_by}, \code{risk_category}, \code{value}. Any column in the input datasets ending with \verb{*rowid} is also passed as is to the output at product level. The exception is any column named exactly \code{rowid}-- which is a reserved name and throws an error. Note this feature makes no sense at company level because potentially multiple rows in the input datasets are summarized into a single row in the output at company level.
}
\description{
\if{html}{\out{<div class="sourceCode {r child=extdata_path("child/intro-sector-profile_upstream.Rmd")}">}}\preformatted{}\if{html}{\out{</div>}}
}
\examples{
library(tiltToyData)
library(readr)
options(readr.show_col_types = FALSE)

companies <- read_csv(toy_sector_profile_upstream_companies())
scenarios <- read_csv(toy_sector_profile_any_scenarios())
inputs <- read_csv(toy_sector_profile_upstream_products())

both <- sector_profile_upstream(companies, scenarios, inputs)
both

both |> unnest_product()

both |> unnest_company()
}
\seealso{
Other main functions: 
\code{\link{emissions_profile_upstream}()},
\code{\link{emissions_profile}()},
\code{\link{sector_profile}()}
}
\concept{main functions}
