% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-test.R
\name{snapshot_accept_private}
\alias{snapshot_accept_private}
\title{Update snapshots in the directory for private tests}
\usage{
snapshot_accept_private(files = NULL)
}
\arguments{
\item{files}{Optionally, filter effects to snapshots from specified files.
This can be a snapshot name (e.g. \code{foo} or \code{foo.md}), a snapshot file name
(e.g. \code{testfile/foo.txt}), or a snapshot file directory (e.g. \verb{testfile/}).}
}
\description{
Update snapshots in the directory for private tests
}
\examples{
\dontrun{
snapshot_accept_private()
}
}
\seealso{
Other developer-oriented functions: 
\code{\link{mvp_path}()},
\code{\link{pctr_path}()},
\code{\link{test_private}()},
\code{\link{use_indicator_issue}()},
\code{\link{wrap_rmd}()}
}
\concept{developer-oriented functions}
