% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilt_path.R
\name{tilt_path}
\alias{tilt_path}
\title{Path to ~/Downloads/tilt}
\usage{
tilt_path(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[fs:path_expand]{fs::path_home}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
A path
}
\description{
Path to ~/Downloads/tilt
}
\examples{
tilt_path()
tilt_path("a", "b")
}
