% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilt_profile.R
\name{tilt_profile}
\alias{tilt_profile}
\title{Create objects of class profile}
\usage{
tilt_profile(data)
}
\arguments{
\item{data}{A data frame with the column \code{companies_id}, and the list columns \code{product} and \code{company} holding the outputs at product and company level.}
}
\value{
An object of class 'profile'.
}
\description{
Create objects of class profile
}
\examples{
library(tibble)
library(tiltIndicator)

product <- company <- tibble(companies_id = 1)
data <- nest_levels(product, company)

class(data)

class(profile(data))

is_tilt_profile(profile(data))
}
\keyword{internal}
