% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiltIndicator-package.R
\docType{package}
\name{tiltIndicator-package}
\alias{tiltIndicator}
\alias{tiltIndicator-package}
\title{tiltIndicator: Indicators for the 'TILT' Project}
\description{
Indicators for the 'TILT' project.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/2DegreesInvesting/tiltIndicator}
  \item Report bugs at \url{https://github.com/2DegreesInvesting/tiltIndicator/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{maurolepore@gmail.com} (\href{https://orcid.org/0000-0002-1986-7988}{ORCID})

Authors:
\itemize{
  \item Tilman Trompke \email{tilman@2degrees-investing.org}
  \item Linda Delacombaz \email{linda@2degrees-investing.org}
  \item Kalash Singhal \email{kalash@2degrees-investing.org}
  \item Lyanne Ho \email{lyho@deloitte.nl}
}

Other contributors:
\itemize{
  \item 2 Degrees Investing Initiative \email{contact@2degrees-investing.org} [copyright holder, funder]
}

}
\keyword{internal}
