% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest.R
\name{unnest_product}
\alias{unnest_product}
\title{Unnest product- and company-level results}
\usage{
unnest_product(data)
}
\arguments{
\item{data}{A nested data frame, e.g. the output of \code{\link[=product_sector]{product_sector()}}.}
}
\value{
A data frame.
}
\description{
Unnest product- and company-level results
}
\examples{
companies <- pstr_companies
scenarios <- xstr_scenarios

both <- product_sector(companies, scenarios)
both |> unnest_product()
both |> unnest_company()
}
\concept{helpers}
