% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_indicator_issue.R
\name{use_indicator_issue}
\alias{use_indicator_issue}
\title{Create a task list that you can paste on a GitHub issue and follow up}
\usage{
use_indicator_issue(indicator = "indicator")
}
\arguments{
\item{indicator}{Character. Hyphen-separated string giving the full name of
the indicator, e.g. "product-carbon-transition-risk".}
}
\value{
Character. A checklist you can paste on a GitHub issue.
}
\description{
Create a task list that you can paste on a GitHub issue and follow up
}
\examples{
use_indicator_issue("product-carbon-transition-risk")
}
\seealso{
Other developer-oriented functions: 
\code{\link{pctr_path}()}
}
\concept{developer-oriented functions}
