% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{use_refactoring_checklist}
\alias{use_refactoring_checklist}
\title{Create a task list that you can paste on a GitHub issue and follow up}
\usage{
use_refactoring_checklist()
}
\value{
Character. A checklist you can paste on a GitHub issue.
}
\description{
Create a task list that you can paste on a GitHub issue and follow up
}
\examples{
use_refactoring_checklist()
}
\keyword{internal}
