% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pstr.R
\docType{data}
\name{weo_2022}
\alias{weo_2022}
\title{WEO 2022}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8 rows and 10 columns.
}
\usage{
weo_2022
}
\description{
WEO 2022
}
\examples{
weo_2022
}
\keyword{datasets}
