% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_rmd.R
\name{wrap_rmd}
\alias{wrap_rmd}
\title{Given an .Rmd file returns a list of it's .csv outputs}
\usage{
wrap_rmd(path)
}
\arguments{
\item{path}{Character. Path to an .Rmd file.}
}
\value{
A named list where each element is one .csv output at the working
directory of the .Rmd file.
}
\description{
This function helps in the early steps of refactoring code from an .Rmd file
into a function. With relatively save, minimal changes to the .Rmd file you
can generate testable output, that you can later use to do more risky
refactoring steps.
}
\examples{
rmd <- system.file("extdata/mtcars.Rmd", package = "tiltIndicator")
wrap_rmd(rmd)
}
\seealso{
Other developer-oriented functions: 
\code{\link{mvp_path}()},
\code{\link{pctr_path}()},
\code{\link{use_indicator_issue}()}
}
\concept{developer-oriented functions}
