% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xctr.R, R/xctr_at_product_level.R
\name{xctr}
\alias{xctr}
\alias{xctr_at_company_level}
\alias{xctr_at_product_level}
\title{Calculate input (or product) carbon transition risk}
\usage{
xctr(companies, co2, low_threshold = 1/3, high_threshold = 2/3)

xctr_at_company_level(data)

xctr_at_product_level(
  companies,
  co2,
  low_threshold = 1/3,
  high_threshold = 2/3
)
}
\arguments{
\item{companies}{A dataframe like \link{companies}.}

\item{co2}{A dataframe like \link{inputs}.}

\item{low_threshold}{A numeric value to segment low and medium transition
risk products.}

\item{high_threshold}{A numeric value to segment medium and high transition
risk products.}

\item{data}{A dataframe. The output at product level.}
}
\value{
At product level, a dataframe with at least columns companies_id, grouped_by, risk_category. At company level, a dataframe with at least columns companies_id, grouped_by, risk_category, value.
}
\description{
These functions calculate the input (or product) carbon transition risk. The
process is the same. What varies is the \code{co2} dataset.
}
\details{
\subsection{Deprecated}{

The \verb{ictr*()} and \verb{pctr*()} functions are now deprecated. Use the \verb{xctr*()}
functions instead.
}

\subsection{Input carbon transition risk (ICTR)}{

The Input Product Carbon Transition Risk Indicator assesses the transition risk
of the input products due to their relative carbon footprint to other input
products. As a default option, each input product is compared to the carbon
footprint of every other input product. Input products with a higher carbon
footprint face a higher risk.

The Input Product Carbon Transition Risk Indicator is therefore similar to the
Product Carbon Transition Risk Indicator, but it focuses on the input products
and not the product of the company. Input products are, for example, resources,
packaging materials, energy and enabling services (such as tractor use on farm)
to produce the product.

After identifying each carbon footprint for one input product, the input
products are ranked according to their footprint. The ranking method is explained
in the \href{https://2degreesinvesting.github.io/tiltIndicator/articles/tiltIndicator.html#thresholds}{Thresholds} section.

After assessing the input products' transition risk based on the carbon
footprint of each product, they are aggregated at the company-level. We derive
what percentage of the input products are high, medium and low transition risk.

The transition risk indicator consists of 2 broad steps:
\enumerate{
\item Score input products: Identifying the input products for each product, and
calculating the relative carbon footprint per input product.
\item Score companies: Aggregating on the company-level.
}

The sample data set includes inputs and co2 footprints for each product from
Ecoinvent and sectors from Europages. NOTE: the following columns are a
completely random selection and do not reflect the true information:
\itemize{
\item co2 footprints (not allowed to share licensed data right now)
\item sectors (as the matching with ecoinvent is not done yet, we do not have one sector per product yet)
}
}

\subsection{Product carbon transition risk (PCTR)}{

The Product Carbon Transition Risk measures the transition risk on the
product-level of a company. The indicator is expressed in the percentage of all
products that are at high risk, medium risk or low risk due to the products'
relative carbon footprint. The assessment is first performed on a product-level
and can then be aggregated to the company-level.

The Product Carbon Transition Risk measures the relative carbon footprint per
product. As a default option each product is compared to the carbon footprint of
every other product. Products with a higher carbon-footprint face a higher risk.

After identifying each carbon footprint for one product, the products are ranked
according to their carbon footprint. The ranking method is explained in the
\href{https://2degreesinvesting.github.io/tiltIndicator/articles/tiltIndicator.html#thresholds}{Thresholds} section.

After categorization, we aggregate all products from the same category and set
them in relation to all products that the company produces. We derive the
Product Carbon Transition Risks.

Please note that carbon footprints, and emissions are used equivalently. Carbon
footprint refers only on the emissions which occur at the production stage of
the product and not the emissions from the inputs. The unit is CO2e in kg.

This indicator provides the share of products with "low", "medium", or "high"
relative production emissions per company.

The output of this indicator contains the following:
\itemize{
\item a column with production emissions
\item a column indicating the percentile relative to (i) all products with same unit (ii) products in same sector (iii) products in same segment
\item a column indicating whether the product has "low", "medium" or "high" relative production emissions.
}
}
}
\examples{
companies <- companies

# ICTR
inputs <- inputs

companies |>
  xctr_at_product_level(inputs)

companies |>
  xctr_at_product_level(inputs) |>
  xctr_at_company_level()

# Same
both <- xctr(companies, inputs)

both |> unnest_product()

both |> unnest_company()

# PCTR
products <- products

companies |>
  xctr_at_product_level(products)

companies |>
  xctr_at_product_level(products) |>
  xctr_at_company_level()

# Same
both <- xctr(companies, products)

both |> unnest_product()

both |> unnest_company()
}
\concept{XCTR functions}
