% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xstr_pivot_type_sector_subsector.R
\name{xstr_pivot_type_sector_subsector}
\alias{xstr_pivot_type_sector_subsector}
\title{Given a messy PSTR \code{companies} dataframe returns a cleaner one}
\usage{
xstr_pivot_type_sector_subsector(data)
}
\arguments{
\item{data}{A dataframe with these columns:
\itemize{
\item ipr_sector
\item ipr_subsector
\item weo_product
\item weo_flow
}}
}
\value{
A \code{companies} dataset as required by PSTR functions.
}
\description{
Given a messy PSTR \code{companies} dataframe returns a cleaner one
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(readr, warn.conflicts = FALSE)

raw_companies <- read_csv(extdata_path("pstr_companies.csv"))
glimpse(raw_companies)

companies <- xstr_pivot_type_sector_subsector(raw_companies)
companies
}
\seealso{
Other helpers: 
\code{\link{pstr_polish_output_at_copmany_level}()},
\code{\link{pstr_prepare_scenario}()},
\code{\link{xstr_prune_companies}()}
}
\concept{helpers}
