% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xstr_polish_output_at_copmany_level.R
\name{xstr_polish_output_at_company_level}
\alias{xstr_polish_output_at_company_level}
\title{Polish output at company level}
\usage{
xstr_polish_output_at_company_level(data)
}
\arguments{
\item{data}{The output of \verb{xstr*()} functions.}
}
\value{
A dataframe.
}
\description{
Polish output at company level
}
\examples{
library(tiltToyData)
library(readr)
options(readr.show_col_types = FALSE)

toy_files()

companies <- toy_path("sector_profile_companies.csv.gz") |>
  read_csv()
sector_profile(companies, xstr_scenarios) |>
  unnest_company() |>
  xstr_polish_output_at_company_level()

sector_profile_upstream(istr_companies, xstr_scenarios, istr_inputs) |>
  unnest_company() |>
  xstr_polish_output_at_company_level()
}
\concept{post-processing helpers}
