% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xstr_polish_output_at_copmany_level.R
\name{xstr_polish_output_at_copmany_level}
\alias{xstr_polish_output_at_copmany_level}
\title{Polish output at company level}
\usage{
xstr_polish_output_at_copmany_level(data)
}
\arguments{
\item{data}{The output of \verb{xstr*()} functions.}
}
\value{
A dataframe.
}
\description{
Polish output at company level
}
\examples{
out <- pstr(pstr_companies, pstr_scenarios)
out |> xstr_polish_output_at_copmany_level()

out <- istr(istr_companies, istr_scenarios, istr_inputs)
out |> xstr_polish_output_at_copmany_level()
}
\seealso{
Other helpers: 
\code{\link{pstr_prepare_scenario}()},
\code{\link{xstr_pivot_type_sector_subsector}()},
\code{\link{xstr_prune_companies}()}
}
\concept{helpers}
