% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xstr_scenarios}
\alias{xstr_scenarios}
\title{XSTR datasets}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 388 rows and 8 columns.
}
\usage{
xstr_scenarios
}
\description{
XSTR datasets
}
\examples{
xstr_scenarios
}
\author{
Tilman Trompke.
}
\concept{XSTR datasets}
\keyword{datasets}
