% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/companies.R
\docType{data}
\name{companies}
\alias{companies}
\title{Fake data of some companies and their products}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10 rows and 5 columns.
}
\usage{
companies
}
\description{
Fake data of some companies and their products
}
\concept{datasets used in the product-sector-transition-risk indicator.
companies}
\keyword{datasets}
