% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/companies.R
\docType{data}
\name{companies}
\alias{companies}
\title{A dataset with fake companies, products, sectors and subsectors}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10 rows and 5 columns.
}
\usage{
companies
}
\description{
\itemize{
\item TODO: Does it have any relationship with tiltData?
\item FIXME: Rename to match existing tiltData, e.g. \code{company_id} -> \code{companies_id}
}
}
\examples{
companies
}
\seealso{
Other datasets for product sector transition risk.: 
\code{\link{ep_weo}},
\code{\link{weo_2022}}
}
\concept{datasets for product sector transition risk.}
\keyword{datasets}
