% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/istr_at_product_level.R,
%   R/pstr_at_product_level.R, R/xctr_at_product_level.R
\name{deprecated}
\alias{istr_at_product_level}
\alias{pstr_at_product_level}
\alias{xctr_at_product_level}
\title{Deprecated}
\usage{
istr_at_product_level(
  companies,
  scenarios,
  inputs,
  low_threshold = ifelse(scenarios$year == 2030, 1/9, 1/3),
  high_threshold = ifelse(scenarios$year == 2030, 2/9, 2/3)
)

pstr_at_product_level(
  companies,
  scenarios,
  low_threshold = ifelse(scenarios$year == 2030, 1/9, 1/3),
  high_threshold = ifelse(scenarios$year == 2030, 2/9, 2/3)
)

xctr_at_product_level(
  companies,
  co2,
  low_threshold = 1/3,
  high_threshold = 2/3
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

All functions of the form \code{f_at_product_level()} or \code{f_at_company_level()}
are now deprecated in favour of a function of the form \code{f()}.
}
\keyword{internal}
