% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\docType{data}
\name{deprecated}
\alias{istr_at_product_level}
\alias{pstr_at_product_level}
\alias{xctr_at_product_level}
\alias{istr_at_company_level}
\alias{pstr_at_company_level}
\alias{xctr_at_company_level}
\alias{xstr_pivot_type_sector_subsector}
\alias{xstr_prepare_scenario}
\alias{xstr_prune_companies}
\alias{xstr_polish_output_at_company_level}
\alias{companies}
\alias{inputs}
\alias{products}
\alias{istr_companies}
\alias{istr_inputs}
\alias{pstr_companies}
\alias{xstr_scenarios}
\title{Deprecated}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 76 rows and 7 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 16 rows and 11 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 3 rows and 8 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 76 rows and 5 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 32 rows and 10 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 152 rows and 10 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 86 rows and 8 columns.
}
\usage{
istr_at_product_level(
  companies,
  scenarios,
  inputs,
  low_threshold = ifelse(scenarios$year == 2030, 1/9, 1/3),
  high_threshold = ifelse(scenarios$year == 2030, 2/9, 2/3)
)

pstr_at_product_level(
  companies,
  scenarios,
  low_threshold = ifelse(scenarios$year == 2030, 1/9, 1/3),
  high_threshold = ifelse(scenarios$year == 2030, 2/9, 2/3)
)

xctr_at_product_level(
  companies,
  co2,
  low_threshold = 1/3,
  high_threshold = 2/3
)

istr_at_company_level(data)

pstr_at_company_level(data)

xctr_at_company_level(data)

xstr_pivot_type_sector_subsector(data)

xstr_prepare_scenario(scenarios)

xstr_prune_companies(data)

xstr_polish_output_at_company_level(data)

companies

inputs

products

istr_companies

istr_inputs

pstr_companies

xstr_scenarios
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

All functions of the form \code{f_at_product_level()} and \code{f_at_company_level()}
are now deprecated in favor of higher-level wrappers (see
\code{?tiltIndicator::rename}).
}
\keyword{datasets}
\keyword{internal}
