% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\docType{data}
\name{deprecated}
\alias{istr_at_product_level}
\alias{pstr_at_product_level}
\alias{xctr_at_product_level}
\alias{istr_at_company_level}
\alias{pstr_at_company_level}
\alias{xctr_at_company_level}
\alias{companies}
\alias{inputs}
\alias{products}
\alias{istr_companies}
\alias{istr_inputs}
\alias{pstr_companies}
\title{Deprecated}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 9 rows and 4 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 33 rows and 6 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 5 rows and 6 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 8 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 74 rows and 10 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 28 rows and 10 columns.
}
\usage{
istr_at_product_level(
  companies,
  scenarios,
  inputs,
  low_threshold = ifelse(scenarios$year == 2030, 1/9, 1/3),
  high_threshold = ifelse(scenarios$year == 2030, 2/9, 2/3)
)

pstr_at_product_level(
  companies,
  scenarios,
  low_threshold = ifelse(scenarios$year == 2030, 1/9, 1/3),
  high_threshold = ifelse(scenarios$year == 2030, 2/9, 2/3)
)

xctr_at_product_level(
  companies,
  co2,
  low_threshold = 1/3,
  high_threshold = 2/3
)

istr_at_company_level(data)

pstr_at_company_level(data)

xctr_at_company_level(data)

companies

inputs

products

istr_companies

istr_inputs

pstr_companies
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

All functions of the form \code{f_at_product_level()} and \code{f_at_company_level()}
are now deprecated in favor of higher-level wrappers (see
\code{?tiltIndicator::rename}).
}
\keyword{datasets}
\keyword{internal}
