% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_default_value.R, R/tilt_profile.R
\name{document_default_value}
\alias{document_default_value}
\alias{document_tilt_profile}
\title{Document the default return value}
\usage{
document_default_value()

document_tilt_profile()
}
\description{
Document the default return value
}
\examples{
document_tilt_profile()

document_default_value()
}
\keyword{internal}
