% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissions_profile.R
\name{emissions_profile}
\alias{emissions_profile}
\title{Calculate the indicator "emissions profile"}
\usage{
emissions_profile(companies, co2, low_threshold = 1/3, high_threshold = 2/3)
}
\arguments{
\item{companies, co2}{A dataframe like the dataset with a matching name in tiltToyData (see \href{https://2degreesinvesting.github.io/tiltToyData/reference/index.html}{Reference}).}

\item{low_threshold}{A numeric value to segment low and medium transition
risk products.}

\item{high_threshold}{A numeric value to segment medium and high transition
risk products.}
}
\value{
A data frame with the column \code{companies_id}, and the nested columns\code{product} and \code{company} holding the outputs at product and company level. Unnesting \code{product} yields a data frame with at least columns \code{companies_id}, \code{grouped_by}, \code{risk_category}. Unnesting \code{company} yields a data frame with at least columns \code{companies_id}, \code{grouped_by}, \code{risk_category}, \code{value}. Any column in the input datasets ending with \verb{*rowid} is also passed as is to the output at product level. The exception is any column named exactly \code{rowid}-- which is a reserved name and throws an error. Note this feature makes no sense at company level because potentially multiple rows in the input datasets are summarized into a single row in the output at company level.
}
\description{
The "emissions profile" measures the transition risk on the product-level of a
company. The indicator is expressed in the percentage of all products that are
at high risk, medium risk or low risk due to the products' relative carbon
footprint. The assessment is first performed on a product-level and can then be
aggregated to the company-level.

The "emissions profile" measures the relative carbon footprint per product. As a
default option each product is compared to the carbon footprint of every other
product. Products with a higher carbon-footprint face a higher risk.

After identifying each carbon footprint for one product, the products are ranked
according to their carbon footprint. The ranking method is explained in the
\href{https://2degreesinvesting.github.io/tiltIndicator/articles/tiltIndicator.html#thresholds}{Thresholds}
section.

After categorization, we aggregate all products from the same category and set
them in relation to all products that the company produces. We derive the
"emissions profile".

Please note that carbon footprints, and emissions are used equivalently. Carbon
footprint refers only on the emissions which occur at the production stage of
the product and not the emissions from the inputs. The unit is CO2e in kg.

This indicator provides the share of products with "low", "medium", or "high"
relative production emissions per company.

The output of this indicator contains the following:
\itemize{
\item a column with production emissions
\item a column indicating the percentile relative to (i) all products with same unit
(ii) products in same sector (iii) products in same segment
\item a column indicating whether the product has "low", "medium" or "high" relative
production emissions.
}
}
\examples{
library(tiltToyData)
library(readr)
options(readr.show_col_types = FALSE)

companies <- read_csv(toy_emissions_profile_any_companies())
products <- read_csv(toy_emissions_profile_products())

both <- emissions_profile(companies, products)
both

both |> unnest_product()

both |> unnest_company()
}
\seealso{
Other main functions: 
\code{\link{emissions_profile_upstream}()},
\code{\link{sector_profile_upstream}()},
\code{\link{sector_profile}()}
}
\concept{main functions}
