% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissions_profile_any_add_profile_ranking.R
\name{emissions_profile_any_compute_profile_ranking}
\alias{emissions_profile_any_compute_profile_ranking}
\title{Add values to categorize}
\usage{
emissions_profile_any_compute_profile_ranking(data)
}
\arguments{
\item{data}{A "co2-like" data frame -- i.e. containing products or
upstream-products (a.k.a. inputs).}
}
\value{
The input data frame with the additional columns \code{grouped_by} and
\code{profile_ranking} and one row per benchmark per company.
}
\description{
Add values to categorize
}
\examples{
library(tiltToyData)
library(readr, warn.conflicts = FALSE)
options(readr.show_col_types = FALSE)

companies <- read_csv(toy_emissions_profile_any_companies())

products <- read_csv(toy_emissions_profile_products())
products |> emissions_profile_any_compute_profile_ranking()

inputs <- read_csv(toy_emissions_profile_upstream_products())
inputs |> emissions_profile_any_compute_profile_ranking()
}
\seealso{
Other pre-processing helpers: 
\code{\link{sector_profile_any_pivot_type_sector_subsector}()},
\code{\link{sector_profile_any_prepare_scenario}()},
\code{\link{sector_profile_any_prune_companies}()}
}
\concept{pre-processing helpers}
