% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep_weo.R
\docType{data}
\name{ep_weo}
\alias{ep_weo}
\title{Mapper between europages and WEO sectors}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 133 rows and 6 columns.
}
\usage{
ep_weo
}
\description{
Mapper between europages and WEO sectors
}
\examples{
ep_weo
}
\seealso{
Other datasets used in the product-sector-transition-risk indicator.: 
\code{\link{companies}},
\code{\link{weo_2022}}
}
\concept{datasets used in the product-sector-transition-risk indicator.}
\keyword{datasets}
