% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstr_data.R
\docType{data}
\name{ep_weo}
\alias{ep_weo}
\title{A bridge between sector and subsectors in europages and in \code{weo_2022}.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 133 rows and 6 columns.
}
\usage{
ep_weo
}
\description{
\itemize{
\item TODO: Where does it come from and what wrangling did you do?
\item FIXME: Rename to match existing tiltData and to use lowercase. E.g. removing
the prefix "ep_" makes it more compatible with tiltData and makes it easier
to join.
}
}
\examples{
ep_weo
}
\seealso{
Other PSTR datasets: 
\code{\link{companies}},
\code{\link{weo_2022}}
}
\concept{PSTR datasets}
\keyword{datasets}
