% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw-data.R
\name{example_raw_companies}
\alias{example_raw_companies}
\title{Example raw companies}
\usage{
example_raw_companies()
}
\value{
A dataframe
}
\description{
Example raw companies
}
\examples{
example_raw_companies()
}
\keyword{internal}
