% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ictr.R
\name{ictr}
\alias{ictr}
\title{Calculate the ICTR indicator}
\usage{
ictr(companies, inputs)
}
\arguments{
\item{companies}{A \link{data.frame} like \link{ictr_companies}.}

\item{inputs}{A \link{data.frame} like \link{ictr_inputs}}
}
\value{
A dataframe with columns \code{companies_id}, \code{transition_risk}, and
scores.
}
\description{
The Input Product Carbon Transition Risk Indicator assesses the transition risk
of the input products due to their relative carbon footprint to other input
products. As a default option, each input product is compared to the carbon
footprint of every other input product. Alternatively, users can also choose to
compare those input products of the same type, so e.g., comparing only raw
materials with each other. Input products with a higher carbon footprint face a
higher risk.

The Input Product Carbon Transition Risk Indicator is therefore similar to the
Product Carbon Transition Risk Indicator, but it focuses on the input products
and not the product of the company. Input products are, for example, resources,
packaging materials, energy and enabling services (such as tractor use on farm)
to produce the product.

After identifying each carbon footprint for one input product, the input
products are ranked according to their footprint. Input products in the highest
percentile (≥70\%) are classified as high transition risk input products. Input
products in the medium percentile (between ≥30\% and <70\%) are classified as
medium transition risk input products. Products in the lowest percentile (<30\%)
are classified as low transition risk input products.

After assessing the input products' transition risk based on the carbon
footprint of each product, they are aggregated at the company-level. We derive
what percentage of the input products are high, medium and low transition risk.

The transition risk indicator consists of 2 broad steps:
\enumerate{
\item Score input products: Identifying the input products for each product, and
calculating the relative carbon footprint per input product.
\item Score companies: Aggregating on the company-level.
}

The sample data set includes inputs and co2 footprints for each product from
Ecoinvent and sectors from Europages. NOTE: the following columns are a
completely random selection and do not reflect the true information:
\itemize{
\item co2 footprints (not allowed to share licensed data right now)
\item sectors (as the matching with ecoinvent is not done yet, we do not have one sector per product yet)
}
}
\examples{
ictr(ictr_companies, ictr_inputs)
}
\seealso{
Other ICTR functions: 
\code{\link{ictr_score_companies}()},
\code{\link{ictr_score_inputs}()}
}
\concept{ICTR functions}
