% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ictr_at_company_level.R
\name{ictr_at_company_level}
\alias{ictr_at_company_level}
\title{Aggregate the input products' scores for each company}
\usage{
ictr_at_company_level(ecoinvent_scores, companies)
}
\arguments{
\item{ecoinvent_scores}{A \link{data.frame}. The output of \link{ictr_at_product_level}.}

\item{companies}{A \link{data.frame} like \link{ictr_companies}.}
}
\value{
A \link{data.frame}.
}
\description{
Calculates on a company-level the percentage of input products that are in
low / medium / high transition risk due to the input products' relative
carbon footprint.

Activities are mapped with the companies' input products using common columns
present in dataframes \code{ecoinvent_scores} and \code{companies}.
}
\examples{
ictr_inputs |>
  ictr_at_product_level() |>
  ictr_at_company_level(ictr_companies)
}
\seealso{
Other internal-ish functions: 
\code{\link{ictr_at_product_level}()}
}
\author{
Kalash Singhal.
}
\concept{internal-ish functions}
