% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ictr_data.R
\docType{data}
\name{ictr_companies}
\alias{ictr_companies}
\title{TODO: Document}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 25 rows and 4 columns.
}
\source{
TODO
}
\usage{
ictr_companies
}
\description{
TODO: Document
}
\examples{
ictr_companies
}
\seealso{
Other ICTR datasets: 
\code{\link{ictr_inputs}}
}
\author{
Kalash Singhal.
}
\concept{ICTR datasets}
\keyword{datasets}
