% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ictr_data.R
\docType{data}
\name{ictr_inputs}
\alias{ictr_inputs}
\title{TODO: Document}
\format{
An object of class \code{data.frame} with 75 rows and 10 columns.
}
\usage{
ictr_inputs
}
\description{
TODO: Document
}
\examples{
ictr_inputs
}
\seealso{
Other ICTR datasets: 
\code{\link{ictr_companies}}
}
\concept{ICTR datasets}
\keyword{datasets}
