% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ictr_data.R
\docType{data}
\name{ictr_inputs}
\alias{ictr_inputs}
\title{A dataset with fake input Co2-footprint, input sector, and activity_uuid.}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 33 rows and 8 columns.
}
\source{
TODO
}
\usage{
ictr_inputs
}
\description{
A dataset with fake input Co2-footprint, input sector, and activity_uuid.
}
\examples{
ictr_inputs
}
\seealso{
Other ICTR datasets: 
\code{\link{ictr_companies}}
}
\author{
Kalash Singhal.
}
\concept{ICTR datasets}
\keyword{datasets}
