% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ictr_score_companies.R
\name{ictr_score_companies}
\alias{ictr_score_companies}
\title{TODO}
\usage{
ictr_score_companies(ecoinvent_scores, companies)
}
\arguments{
\item{ecoinvent_scores}{TODO}

\item{companies}{TODO}
}
\value{
A \link{data.frame} with three rows for each company and these columns:
TODO
}
\description{
TODO
}
\examples{
library(dplyr, warn.conflicts = FALSE)

one_company <- ictr_companies |> filter(company_id \%in\% first(company_id))
one_company

ictr_inputs |>
  ictr_score_inputs() |>
  ictr_score_companies(one_company)
}
\seealso{
Other ICTR functions: 
\code{\link{ictr_score_inputs}()}
}
\author{
Kalash Singhal.
}
\concept{ICTR functions}
