% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ictr_score_inputs.R
\name{ictr_score_inputs}
\alias{ictr_score_inputs}
\title{TODO}
\usage{
ictr_score_inputs(inputs, low_threshold = 0.3, high_threshold = 0.7)
}
\arguments{
\item{inputs}{TODO}

\item{low_threshold}{TODO}

\item{high_threshold}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
ictr_score_inputs(ictr_inputs)
}
\seealso{
Other ICTR functions: 
\code{\link{ictr_score_companies}()}
}
\concept{ICTR functions}
