% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ictr_score_inputs.R
\name{ictr_score_inputs}
\alias{ictr_score_inputs}
\title{Aggregate rank and score for each input product}
\usage{
ictr_score_inputs(inputs, low_threshold = 0.3, high_threshold = 0.7)
}
\arguments{
\item{inputs}{A \link{data.frame} like \link{ictr_inputs}}

\item{low_threshold}{A numeric value to segment low and medium transition
risk products.}

\item{high_threshold}{A numeric value to segment medium and high transition
risk products.}
}
\value{
A \link{data.frame} with columns:
\itemize{
\item All the columns from the \code{inputs} dataset.
\item New columns:
\itemize{
\item \code{perc_all}
\item \code{perc_unit}
\item \code{perc_sec}
\item \code{perc_unit_sec}
\item \code{score_all}
\item \code{score_unit}
\item \code{score_sector}
\item \code{score_unit_sec}
}
}
}
\description{
Input products will be ranked according to their footprint. Input products in
the highest percentile (≥70\%) will be classified as high transition risk
input products. Input products in the medium percentile (between ≥30\% and
<70\%) will be classified as medium transition risk input products. Products
in the lowest percentile (<30\%) will be classified as low transition risk
input products.

Therefore, each input product will receive either a score of 'high',
'medium', or 'low' based on the above mentioned classification criteria.
}
\examples{
library(dplyr, warn.conflicts = FALSE)

# Minimum dataset
crucial <- c("input_co2", "input_sector", "unit")
data <- ictr_inputs |>
  slice(1L) |>
  select(all_of(crucial))
data

ictr_score_inputs(data) |> glimpse()

# With fewer columns returns an error
data2 <- data |> select(-unit)
try(ictr_score_inputs(data2))

# Additional columns are added to the output
data3 <- data |> mutate(new = 1)
ictr_score_inputs(data3) |> relocate(new)

# You may customize thresholds
data4 <- ictr_inputs |>
  slice(1:3) |>
  select(all_of(crucial))

# Default
ictr_score_inputs(data4) |>
  relocate(starts_with("score_"))

# Custom
ictr_score_inputs(data4, low_threshold = 1, high_threshold = 2) |>
  relocate(starts_with("score_"))
}
\seealso{
Other ICTR functions: 
\code{\link{ictr_score_companies}()},
\code{\link{ictr}()}
}
\author{
Kalash Singhal.
}
\concept{ICTR functions}
