% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ictr_score_inputs.R
\name{ictr_score_inputs}
\alias{ictr_score_inputs}
\title{TODO}
\usage{
ictr_score_inputs(inputs, low_threshold = 0.3, high_threshold = 0.7)
}
\arguments{
\item{inputs}{TODO}

\item{low_threshold}{TODO}

\item{high_threshold}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
library(dplyr, warn.conflicts = FALSE)

# Minimum dataset
crucial <- c("input_co2", "input_sector", "unit")
data <- ictr_inputs |>
  slice(1L) |>
  select(all_of(crucial))
data

ictr_score_inputs(data) |> glimpse()

# With fewer columns returns an error
data2 <- data |> select(-unit)
try(ictr_score_inputs(data2))

# Additional columns are added to the output
data3 <- data |> mutate(new = 1)
ictr_score_inputs(data3) |> relocate(new)

# You may customize thresholds
data4 <- ictr_inputs |>
  slice(1:3) |>
  select(all_of(crucial))

# Default
ictr_score_inputs(data4) |>
  relocate(starts_with("score_"))

# Custom
ictr_score_inputs(data4, low_threshold = 1, high_threshold = 2) |>
  relocate(starts_with("score_"))
}
\seealso{
Other ICTR functions: 
\code{\link{ictr_score_companies}()}
}
\concept{ICTR functions}
