% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/istr_companies.R
\docType{data}
\name{istr_companies}
\alias{istr_companies}
\title{TODO}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 33 rows and 36 columns.
}
\usage{
istr_companies
}
\description{
TODO
}
\examples{
istr_companies
}
\concept{ISTR datasets}
\keyword{datasets}
