% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{istr_datasets}
\alias{istr_datasets}
\alias{istr_companies}
\alias{xstr_scenarios}
\alias{istr_inputs}
\title{ISTR datasets}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 8 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 388 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 74 rows and 10 columns.
}
\usage{
istr_companies

xstr_scenarios

istr_inputs
}
\description{
ISTR datasets
}
\examples{
istr_companies

xstr_scenarios

istr_inputs
}
\author{
Lyanne Ho.
}
\concept{ISTR datasets}
\keyword{datasets}
