% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jitter_range.R
\name{jitter_range}
\alias{jitter_range}
\title{Expand a range (towards the left and right) adding an amount of random noise}
\usage{
jitter_range(data, amount = 0.1)
}
\arguments{
\item{data}{A dataframe with columns \code{min} and \code{max}.}

\item{amount}{Numeric. A single value giving the amount of jitter.}
}
\value{
The input dataframe with the additional columns \code{min_jitter} and
\code{max_jitter}.
}
\description{
Expand a range (towards the left and right) adding an amount of random noise
}
\examples{
library(tibble)

data <- tibble(min = 1:2, max = 3:4)

data |> jitter_range()

data |> jitter_range(amount = 0.9)
}
\seealso{
Other post-processing helpers: 
\code{\link{sector_profile_any_polish_output_at_company_level}()},
\code{\link{summarize_range}()}
}
\concept{post-processing helpers}
