% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jitter_range.R
\name{jitter_range}
\alias{jitter_range}
\title{Expand a range adding some random noise}
\usage{
jitter_range(data, factor = 1, amount = NULL)
}
\arguments{
\item{data}{A dataframe with columns \code{min} and \code{max}.}

\item{factor}{numeric.}

\item{amount}{numeric; if positive, used as \emph{amount} (see below),
    otherwise, if \code{= 0} the default is \code{factor * z/50}.

    Default (\code{NULL}): \code{factor * d/5} where \code{d} is about
    the smallest difference between \code{x} values.}
}
\value{
The input dataframe with the additional columns \code{min_jitter} and
\code{max_jitter}.
}
\description{
This function expands a range by adding noise to the left of the minimum
values and to the right of the maximum values.
}
\examples{
library(tibble)
set.seed(123)

data <- tibble(min = -2:2, max = -1:3)

data |> jitter_range(amount = 0.1)

data |> jitter_range(amount = 2)
}
\seealso{
\code{\link[=jitter]{jitter()}}

Other helpers: 
\code{\link{summarize_range}()},
\code{\link{unnest_product}()}
}
\concept{helpers}
