% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jitter_towards.R
\name{jitter_towards}
\alias{jitter_towards}
\title{Add a small amount of random noise to a vector, towards the left or right}
\usage{
jitter_towards(x, towards = c("left", "right"), amount = 0.1)
}
\arguments{
\item{x}{A numeric vector.}

\item{towards}{Character. Either "left" (default) or "right".}

\item{amount}{Numeric. A single value giving the amount of jitter.}
}
\value{
A numeric vector.
}
\description{
Add a small amount of random noise to a vector, towards the left or right
}
\examples{
1:3 |> jitter_towards("left")

1:3 |> jitter_towards("right")

1:3 |> jitter_towards("right", amount = 0.9)
}
\seealso{
Other post-processing helpers: 
\code{\link{sector_profile_any_polish_output_at_company_level}()},
\code{\link{summarize_range}()}
}
\concept{post-processing helpers}
