% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pctr.R, R/xctr.R
\name{pctr}
\alias{pctr}
\alias{pctr_at_product_level}
\alias{pctr_at_company_level}
\title{Calculate the PCTR indicator}
\usage{
pctr(companies, co2, low_threshold = 0.3, high_threshold = 0.7)

pctr_at_product_level(
  companies,
  co2,
  low_threshold = 0.3,
  high_threshold = 0.7
)

pctr_at_company_level(data)
}
\arguments{
\item{companies}{A dataframe like \link{pctr_companies}.}

\item{co2}{A dataframe like \link{pctr_ecoinvent_co2}.}

\item{low_threshold}{A numeric value to segment low and medium transition
risk products.}

\item{high_threshold}{A numeric value to segment medium and high transition
risk products.}

\item{data}{A dataframe. The output at product level.}
}
\value{
At product level, a dataframe with at least columns companies_id, grouped_by, risk_category. At company level, a dataframe with at least columns companies_id, grouped_by, risk_category, value.
}
\description{
The Product Carbon Transition Risk measures the transition risk on the
product-level of a company. The indicator is expressed in the percentage of all
products that are at high risk, medium risk or low risk due to the products'
relative carbon footprint. The assessment is first performed on a product-level
and can then be aggregated to the company-level.

The Product Carbon Transition Risk measures the relative carbon footprint per
product. As a default option each product is compared to the carbon footprint of
every other product. Products with a higher carbon-footprint face a higher risk.

After identifying each carbon footprint for one product, the products are ranked
according to their carbon footprint. Products in the highest percentile (≥70\%)
are classified as high transition risk products. Products in the medium
percentile (between ≥30\% and <70\%) are classified as medium transition risk
products. Products in the lowest percentile (<30\%) are classified as low
transition risk products.

After categorization, we aggregate all products from the same category and set
them in relation to all products that the company produces. We derive the
Product Carbon Transition Risks.

Please note that carbon footprints, and emissions are used equivalently. Carbon
footprint refers only on the emissions which occur at the production stage of
the product and not the emissions from the inputs. The unit is CO2e in kg.

This indicator provides the share of products with "low", "medium", or "high"
relative production emissions per company.

The output of this indicator contains the following:
\itemize{
\item a column with production emissions
\item a column indicating the percentile relative to (i) all products with same unit (ii) products in same sector (iii) products in same segment
\item a column indicating whether the product has "low", "medium" or "high" relative production emissions.
}
}
\examples{
companies <- pctr_companies
co2 <- pctr_ecoinvent_co2

# Product level
companies |>
  pctr_at_product_level(co2)

# Company level
companies |>
  pctr_at_product_level(co2) |>
  pctr_at_company_level()

# Same
pctr(companies, co2)
}
\concept{PCTR functions}
