% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pctr_at_product_level.R
\name{pctr_at_product_level}
\alias{pctr_at_product_level}
\title{Aggregate rank and score for each activity}
\usage{
pctr_at_product_level(co2, low_threshold = 0.3, high_threshold = 0.7)
}
\arguments{
\item{co2}{A \link{data.frame} like \link{pctr_ecoinvent_co2}.}

\item{low_threshold}{A numeric value to segment low and medium transition
risk products.}

\item{high_threshold}{A numeric value to segment medium and high transition
risk products.}
}
\value{
A \link{data.frame}.
}
\description{
Activities will be ranked according to their carbon footprint. Activities in
the highest percentile (≥70\%) will be classified as high transition risk
products. Activities in the medium percentile (between ≥30\% and <70\%) will be
classified as medium transition risk products. Activities in the lowest
percentile (<30\%) will be classified as low transition risk products.

Therefore, each activity will receive either a score of 'high', 'medium', or
'low' based on the above mentioned classification criteria.
}
\examples{
pctr_at_product_level(pctr_ecoinvent_co2)
}
\seealso{
Other internal-ish functions: 
\code{\link{ictr_at_company_level}()},
\code{\link{ictr_at_product_level}()},
\code{\link{pctr_at_company_level}()}
}
\author{
Tilman Trompke.
}
\concept{internal-ish functions}
