% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pctr_data.R
\docType{data}
\name{pctr_companies}
\alias{pctr_companies}
\title{TODO add title}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 25 rows and 4 columns.
}
\source{
TODO add source
}
\usage{
pctr_companies
}
\description{
TODO add title
}
\examples{
pctr_companies
}
\seealso{
Other PCTR datasets: 
\code{\link{pctr_ecoinvent_co2}}
}
\author{
Tilman Trompke.
}
\concept{PCTR datasets}
\keyword{datasets}
