% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pctr_score_companies.R
\name{pctr_score_companies}
\alias{pctr_score_companies}
\title{Title TODO}
\usage{
pctr_score_companies(scored_activities, companies)
}
\arguments{
\item{scored_activities}{A \link{data.frame}. The output of
\link{pctr_score_activities}.}

\item{companies}{A \link{data.frame} like \link{pctr_companies}.}
}
\value{
A \link{data.frame} with columns ... TODO
}
\description{
Description TODO
}
\details{
Details (optional) TODO
}
\examples{
pctr_ecoinvent_co2 |>
  pctr_score_activities(low_threshold = 0.3, high_threshold = 0.7) |>
  pctr_score_companies(pctr_companies)
}
