% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pctr_score_companies.R
\name{pctr_score_companies}
\alias{pctr_score_companies}
\title{Aggregate the products' scores for each company}
\usage{
pctr_score_companies(scored_activities, companies)
}
\arguments{
\item{scored_activities}{A \link{data.frame}. The output of
\link{pctr_score_activities}.}

\item{companies}{A \link{data.frame} like \link{pctr_companies}.}
}
\value{
A \link{data.frame} with three rows for each company and these columns:
\itemize{
\item \code{company_id}
\item \code{score}
\item \code{share_all}
\item \code{share_unit}
\item \code{share_unit_sec}
where \code{share_all}, \code{share_unit}, and \code{share_unit_sec} holds the
aggregated scores in percentage.
}
}
\description{
Calculates on a company-level the percentage of products that are in low /
medium / high transition risk due to the products’ relative carbon footprint.

Activities are mapped with the companies' products using common columns
present in dataframes \code{scored_activities} and \code{companies}
}
\examples{
library(dplyr, warn.conflicts = FALSE)

first_company <- pctr_companies |> filter(company_id \%in\% first(company_id))
first_company

pctr_ecoinvent_co2 |>
  pctr_score_activities(low_threshold = 0.3, high_threshold = 0.7) |>
  pctr_score_companies(first_company)
}
\seealso{
Other PCTR functions: 
\code{\link{pctr_score_activities}()}
}
\concept{PCTR functions}
