% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstr.R, R/xctr_at_product_level.R
\name{pstr}
\alias{pstr}
\alias{pstr_at_product_level}
\alias{pstr_at_company_level}
\title{Calculate the PSTR indicator}
\usage{
pstr(companies, scenarios, low_threshold = 30, high_threshold = 70)

pstr_at_product_level(
  companies,
  scenarios,
  low_threshold = 30,
  high_threshold = 70
)

pstr_at_company_level(
  companies,
  co2,
  low_threshold = 1/3,
  high_threshold = 2/3
)
}
\arguments{
\item{companies}{A dataframe like \link{pstr_companies}.}

\item{scenarios}{A dataframe like \link{pstr_scenarios}.}

\item{low_threshold}{A numeric value to segment low and medium reduction
targets.}

\item{high_threshold}{A numeric value to segment medium and high reduction
targets.}
}
\value{
At product level, a dataframe with at least columns companies_id, grouped_by, risk_category. At company level, a dataframe with at least columns companies_id, grouped_by, risk_category, value.
}
\description{
The Product Sector Transition Risk Indicator measures the transition risk of
products based on the sector's emissions targets the product belongs to. Those
sector emission reduction targets vary across scenarios (e.g., net zero
scenario or 1.5° scenario) and the time horizon (e.g., reduction needed in
2030, 2040, 2050 to achieve the targets).

After assessing each product, all the products with the same category are
aggregated and set in relation to all products of the company. We, therefore,
derive company-level information.
}
\examples{
companies <- pstr_companies
scenarios <- pstr_scenarios

# Product level
companies |>
  pstr_at_product_level(scenarios)

# Company level
companies |>
  pstr_at_product_level(scenarios) |>
  pstr_at_company_level()

# Same
pstr(companies, scenarios)
}
\concept{PSTR functions}
