% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstr.R
\name{pstr}
\alias{pstr}
\title{Calculate the PSTR indicator}
\usage{
pstr(companies, scenarios)
}
\arguments{
\item{companies}{A dataframe like \link{pstr_companies}.}

\item{scenarios}{A dataframe like \link{pstr_scenarios}.}
}
\value{
At company level, a dataframe with columns \code{companies_id}, \code{grouped_by}, \code{risk_category}, \code{value}.
}
\description{
The Product Sector Transition Risk Indicator measures the transition risk of
products based on the sector's emissions targets the product belongs to. Those
sector emission reduction targets vary across scenarios (e.g., net zero
scenario or 1.5° scenario) and the time horizon (e.g., reduction needed in
2030, 2040, 2050 to achieve the targets).

After assessing each product, all the products with the same category are
aggregated and set in relation to all products of the company. We, therefore,
derive company-level information.
}
\examples{
companies <- pstr_companies
scenarios <- pstr_scenarios

pstr(companies, scenarios)
}
\seealso{
Other PSTR functions: 
\code{\link{pstr_at_company_level}()}
}
\concept{PSTR functions}
