% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstr.R
\name{pstr}
\alias{pstr}
\title{Calculate the PSTR indicator}
\usage{
pstr(
  companies,
  scenarios,
  low_threshold = ifelse(scenarios$year == 2030, 1/9, 1/3),
  high_threshold = ifelse(scenarios$year == 2030, 2/9, 2/3)
)
}
\arguments{
\item{companies}{A dataframe like \link{pstr_companies}.}

\item{scenarios}{A dataframe like \link{xstr_scenarios}.}

\item{low_threshold}{A numeric value to segment low and medium reduction
targets.}

\item{high_threshold}{A numeric value to segment medium and high reduction
targets.}
}
\value{
At product level, a dataframe with at least columns companies_id, grouped_by, risk_category. At company level, a dataframe with at least columns companies_id, grouped_by, risk_category, value.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was deprecated because its name is terse and it doesn't help
access results at product level.

The Product Sector Transition Risk Indicator measures the transition risk of
products based on the sector's emissions targets the product belongs to. Those
sector emission reduction targets vary across scenarios (e.g., net zero
scenario or 1.5° scenario) and the time horizon (e.g., reduction needed in
2030, 2040, 2050 to achieve the targets).

After assessing each product, all the products with the same category are
aggregated and set in relation to all products of the company. We, therefore,
derive company-level information.
}
\examples{
library(dplyr)
library(tidyr)

companies <- pstr_companies
scenarios <- xstr_scenarios

# Deprecated
company_level <- pstr(companies, scenarios)
company_level

# Now
both <- product_sector(companies, scenarios)

company_level <- both |> select(-product) |> unnest(company)
company_level

# Also
product_level <- both |> select(-company) |> unnest(product)
product_level
}
\seealso{
Other PSTR functions: 
\code{\link{product_sector}()}
}
\concept{PSTR functions}
\keyword{internal}
