% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstr.R
\name{pstr}
\alias{pstr}
\title{Calculate the PSTR indicator}
\usage{
pstr(companies, scenario, mapper)
}
\arguments{
\item{companies}{A \link{data.frame} like \link{pstr_companies}.}

\item{scenario}{A dataframe with scenario data, e.g. WEO.}
}
\value{
A dataframe with columns \code{id}, \code{transition_risk}, and scores.
}
\description{
The Product Sector Transition Risk Indicator measures the transition risk of
products based on the sector's emissions targets the product belongs to. Those
sector emission reduction targets vary across scenarios (e.g., net zero
scenario or 1.5° scenario) and the time horizon (e.g., reduction needed in
2030, 2040, 2050 to achieve the targets).

After assessing each product, all the products with the same category are
aggregated and set in relation to all products of the company. We, therefore,
derive company-level information.
}
\examples{
pstr(pstr_companies, pstr_weo_2022, pstr_ep_weo)
}
\seealso{
Other PSTR functions: 
\code{\link{pstr_add_reductions}()},
\code{\link{pstr_add_transition_risk}()},
\code{\link{pstr_aggregate_scores}()}
}
\concept{PSTR functions}
