% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstr_add_transition_risk.R
\name{pstr_add_transition_risk}
\alias{pstr_add_transition_risk}
\title{Categorize sector emission reduction targets}
\usage{
pstr_add_transition_risk(with_reductions)
}
\arguments{
\item{with_reductions}{A data frame. The output of \code{\link[=pstr_add_reductions_old]{pstr_add_reductions_old()}}.}
}
\value{
A data frame with:
\itemize{
\item All the columns from the \code{companies} dataset.
\item All the columns from the \code{ep_weo} dataset, but the columns
\code{EP_sector} and \code{EP_subsector} are named \code{sector} and \code{subsector}
respectively.
\item All the columns from the \code{weo_2022}dataset, but the columns
\code{weo_product_mapper} and \code{weo_flow_mapper} are named \code{product} and
\code{flow} respectively.
\item New column:
\itemize{
\item \code{transition_risk} that holds the categorization of the \code{reductions}
column.
}
}
}
\description{
#TODO : Ask what are the obsoletes reductions values
Translates the \code{reductions} column into three categories:
\itemize{
\item "low" if \code{reductions} <= 30.0
\item "medium" if 30.0 < \code{reductions} <= 70.0
\item "high" if \code{reductions} > 70.0
}
}
\examples{
pstr_companies |>
  pstr_add_reductions_old(pstr_ep_weo, pstr_weo_2022) |>
  pstr_add_transition_risk()
}
\seealso{
Other PSTR functions: 
\code{\link{pstr_add_reductions_old}()},
\code{\link{pstr_aggregate_scores}()}
}
\author{
Linda Delacombaz.
}
\concept{PSTR functions}
