% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstr_aggregate_scores.R
\name{pstr_aggregate_scores}
\alias{pstr_aggregate_scores}
\title{Title}
\usage{
pstr_aggregate_scores(data)
}
\arguments{
\item{data}{A data frame. The ouptput of \code{\link[=pstr_add_transition_risk]{pstr_add_transition_risk()}}.}
}
\value{
TODO
}
\description{
Title
}
\examples{
# TODO
}
