% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstr_aggregate_scores.R
\name{pstr_aggregate_scores}
\alias{pstr_aggregate_scores}
\title{Aggregate the products' scores for each company}
\usage{
pstr_aggregate_scores(with_transition_risk, companies)
}
\arguments{
\item{with_transition_risk}{A \link{data.frame}. The output of
\code{\link[=pstr_add_transition_risk]{pstr_add_transition_risk()}}.}

\item{companies}{A \link{data.frame} like \link{pstr_companies}.}
}
\value{
A \link{data.frame} with the columns:
\itemize{
\item \code{company_name}
\item \code{transition_risk}
\item \code{scenario}
\item \code{year}
\item \code{score_aggregated}, which holds the aggregated scores in percentage.
}
}
\description{
Calculates on a company-level the percentage of products that are in low /
medium / high transition risk.
}
\examples{
pstr_companies |>
  pstr_add_reductions(pstr_ep_weo, pstr_weo_2022) |>
  pstr_add_transition_risk() |>
  pstr_aggregate_scores(pstr_companies)
}
\seealso{
Other PSTR functions: 
\code{\link{pstr_add_reductions}()},
\code{\link{pstr_add_transition_risk}()},
\code{\link{pstr_new_add_reductions}()},
\code{\link{pstr_prepare_companies}()},
\code{\link{pstr_prepare_scenario}()},
\code{\link{pstr}()}
}
\author{
Linda Delacombaz.
}
\concept{PSTR functions}
\keyword{internal}
