% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstr_data.R
\docType{data}
\name{pstr_companies}
\alias{pstr_companies}
\title{A dataset with fake companies, products, sectors and subsectors}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10 rows and 5 columns.
}
\usage{
pstr_companies
}
\description{
\itemize{
\item TODO: Does it have any relationship with tiltData?
\item FIXME: Rename to match existing tiltData, e.g. \code{company_id} -> \code{companies_id}
}
}
\examples{
pstr_companies
}
\seealso{
Other PSTR datasets: 
\code{\link{pstr_ep_weo}},
\code{\link{pstr_weo_2022}}
}
\concept{PSTR datasets}
\keyword{datasets}
