% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstr_new_add_reductions.R
\name{pstr_new_add_reductions}
\alias{pstr_new_add_reductions}
\title{Add the emission reduction targets to the companies dataset}
\usage{
pstr_new_add_reductions(companies, scenarios)
}
\arguments{
\item{companies}{TODO}

\item{scenario}{TODO}
}
\value{
A dataframe with:
\itemize{
\item All the columns from the \code{companies} dataset.
\item New columns:
\itemize{
\item All the columns from the \code{scenario} dataset.
}
}
}
\description{
Adds the emission reduction values for each company's product(s).
}
\examples{
scenarios <- pstr_scenarios
companies <- pstr_companies
pstr_new_add_reductions(companies, scenarios)
}
\seealso{
Other PSTR functions: 
\code{\link{pstr_add_transition_risk}()},
\code{\link{pstr_aggregate_scores}()},
\code{\link{pstr_old_add_reductions}()},
\code{\link{pstr}()}
}
\concept{PSTR functions}
