% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstr_plot_company.R
\name{pstr_plot_company}
\alias{pstr_plot_company}
\title{Title}
\usage{
pstr_plot_company(with_score_aggregated, company_name)
}
\arguments{
\item{data}{A data frame. The ouptput of \code{\link[=pstr_aggregate_scores]{pstr_aggregate_scores()}}.}
}
\value{
TODO
}
\description{
Title
}
\examples{
# TODO
}
