% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstr_prepare_companies.R
\name{pstr_prepare_companies}
\alias{pstr_prepare_companies}
\title{Title}
\usage{
pstr_prepare_companies(data)
}
\arguments{
\item{data}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{
# TODO
}
\seealso{
Other PSTR functions: 
\code{\link{pstr_add_reductions}()},
\code{\link{pstr_add_transition_risk}()},
\code{\link{pstr_aggregate_scores}()},
\code{\link{pstr_new_add_reductions}()},
\code{\link{pstr}()}
}
\concept{PSTR functions}
\keyword{internal}
