% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstr_prepare_companies.R
\name{pstr_prepare_companies}
\alias{pstr_prepare_companies}
\title{Help prepare the PSTR \code{companies} data}
\usage{
pstr_prepare_companies(data)
}
\arguments{
\item{data}{A dataframe with these columns:
\itemize{
\item ipr_sector.x
\item ipr_sector.y
\item ipr_subsector.x
\item ipr_subsector.y
\item weo_product.x
\item weo_product.y
\item weo_flow.x
\item weo_flow.y
}}
}
\value{
A \code{copmanies} dataset as required by PSTR functions.
}
\description{
Help prepare the PSTR \code{companies} data
}
\examples{
library(dplyr, warn.conflicts = FALSE)

companies <- pstr_new_companies |>
  slice(1) |>
  glimpse()

pstr_prepare_companies(companies)
}
\seealso{
Other pre-processing helpers: 
\code{\link{pstr_prepare_scenario}()}
}
\concept{pre-processing helpers}
