% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstr_prepare_scenario.R
\name{pstr_prepare_scenario}
\alias{pstr_prepare_scenario}
\title{Given a named list of scenarios returns a single, cleaner dataframe}
\usage{
pstr_prepare_scenario(scenarios)
}
\arguments{
\item{scenarios}{A named list of identically structured scenarios.}
}
\value{
A single, cleaner dataframe with an additional column to identify
which rows come from which scenario.
}
\description{
Given a named list of scenarios returns a single, cleaner dataframe
}
\examples{
library(dplyr, warn.conflicts = FALSE)

weo <- slice(pstr_new_weo_2022, 1:2)
ipr <- slice(pstr_new_weo_2022, 1:2)

scenarios <- list(weo = weo, ipr = ipr)
pstr_prepare_scenario(scenarios)

# You prepare other scenarios
scenarios <- list(scen1 = weo, scen2 = weo)
pstr_prepare_scenario(scenarios) |>
  relocate(type)
}
\seealso{
Other pre-processing helpers: 
\code{\link{pstr_prepare_companies}()}
}
\concept{pre-processing helpers}
