% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstr_prepare_scenario.R
\name{pstr_prepare_scenario}
\alias{pstr_prepare_scenario}
\title{Given a named list of scenarios returns a cleaner \code{scenarios} dataframe}
\usage{
pstr_prepare_scenario(scenarios)
}
\arguments{
\item{scenarios}{A named list of identically structured scenarios.}
}
\value{
A single, cleaner dataframe with an additional column to identify
which rows come from which scenario.
}
\description{
Given a named list of scenarios returns a cleaner \code{scenarios} dataframe
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(readr, warn.conflicts = FALSE)

raw_weo <- read_csv(extdata_path("pstr_weo_2022.csv")) |>
  rename(weo_sector = "weo_product", weo_subsector = "weo_flow")
raw_ipr <- read_csv(extdata_path("pstr_ipr_2022.csv"))
raw_scenarios <- list(weo = raw_weo, ipr = raw_ipr)

pstr_prepare_scenario(raw_scenarios)
}
\seealso{
Other pre-processing helpers: 
\code{\link{pstr_prepare_companies}()}
}
\concept{pre-processing helpers}
