% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename}
\alias{pstr}
\alias{istr}
\title{Renamed functions and datasets renamed}
\usage{
pstr(
  companies,
  scenarios,
  low_threshold = ifelse(scenarios$year == 2030, 1/9, 1/3),
  high_threshold = ifelse(scenarios$year == 2030, 2/9, 2/3)
)

istr(
  companies,
  scenarios,
  inputs,
  low_threshold = ifelse(scenarios$year == 2030, 1/9, 1/3),
  high_threshold = ifelse(scenarios$year == 2030, 2/9, 2/3)
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Conceptual changes:
\itemize{
\item PCTR becomes Emissions Profile.
\item ICTR becomes Emissions Profile Upstream.
\item PSTR becomes Sector Profile.
\item ISTR becomes Sector Profile Upstream.
}

Motivation:
\itemize{
\item The names are more informative and easier to remember.
\item All indicators now share only one common word (profile) instead of two
(transition risk).
\item The word "upstream" is more familiar for users in banks than "inputs".
\item The word "emissions" replaces "carbon" because the data that we use actually
take into account CO2 equivalents, i.e. all GHG.
\item Compared to the phrase "transition risk", the word "profile" better reflects
that the indicators cannot only be used for risk assessment but also for
other things, such as broader sustainability assessment, engagement,
reporting, etc.
}

Implementation changes:
\itemize{
\item v0.0.0.9084: \code{pstr()} -> \code{sector_profile()}
\item v0.0.0.9085: \code{istr()} -> \code{sector_profile_upstream()}
}
}
\keyword{internal}
