% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename}
\alias{pstr}
\alias{istr}
\alias{xctr}
\title{Renamed functions and datasets}
\usage{
pstr(
  companies,
  scenarios,
  low_threshold = ifelse(scenarios$year == 2030, 1/9, 1/3),
  high_threshold = ifelse(scenarios$year == 2030, 2/9, 2/3)
)

istr(
  companies,
  scenarios,
  inputs,
  low_threshold = ifelse(scenarios$year == 2030, 1/9, 1/3),
  high_threshold = ifelse(scenarios$year == 2030, 2/9, 2/3)
)

xctr(companies, co2, low_threshold = 1/3, high_threshold = 2/3)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Conceptual changes:
\itemize{
\item PCTR becomes Emissions Profile.
\item ICTR becomes Emissions Profile Upstream.
\item PSTR becomes Sector Profile.
\item ISTR becomes Sector Profile Upstream.
}

Motivation:
\itemize{
\item The names are more informative and easier to remember.
\item All indicators now share only one common word (profile) instead of two
(transition risk).
\item The word "upstream" is more familiar for users in banks than "inputs".
\item The word "emissions" replaces "carbon" because the data that we use actually
take into account CO2 equivalents, i.e. all green house gasses.
\item Compared to the phrase "transition risk", the word "tilt_profile" better reflects
that the indicators cannot only be used for risk assessment but also for
other things, such as broader sustainability assessment, engagement,
reporting, etc.
}

Implementation changes:
\itemize{
\item v0.0.0.9084: \code{pstr()} -> \code{sector_profile()}
\item v0.0.0.9085: \code{istr()} -> \code{sector_profile_upstream()}
\item v0.0.0.9086: \code{xctr(companies, products)} -> \code{emissions_profile()}
\item v0.0.0.9087: \code{xctr(companies, inputs)} -> \code{emissions_profile_upstream()}
\item v0.0.0.9089: All datasets names match the functions above and moved to \href{https://2degreesinvesting.github.io/tiltToyData/reference/index.html}{tiltToyData}
\item v0.0.0.9092: \code{xstr_pivot_type_sector_subsector()} -> \code{sector_profile_any_pivot_type_sector_subsector()}
\item v0.0.0.9092: \code{xstr_prepare_scenario()} -> \code{sector_profile_any_prepare_scenario()}
\item v0.0.0.9092: \code{xstr_prune_companies()} -> \code{sector_profile_any_prune_companies()}
\item v0.0.0.9092: \code{xstr_polish_output_at_company_level()} -> \code{sector_profile_any_polish_output_at_company_level()}
}
}
\keyword{internal}
