% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sector_profile.R
\name{sector_profile}
\alias{sector_profile}
\title{Calculate the indicator "sector profile"}
\usage{
sector_profile(
  companies,
  scenarios,
  low_threshold = ifelse(scenarios$year == 2030, 1/9, 1/3),
  high_threshold = ifelse(scenarios$year == 2030, 2/9, 2/3)
)
}
\arguments{
\item{companies, scenarios}{A dataframe like the dataset with a matching name in tiltToyData (see \href{https://2degreesinvesting.github.io/tiltToyData/reference/index.html}{Reference}).}

\item{low_threshold}{A numeric value to segment low and medium reduction
targets.}

\item{high_threshold}{A numeric value to segment medium and high reduction
targets.}
}
\value{
A data frame with the column \code{companies_id}, and the nested columns\code{product} and \code{company} holding the outputs at product and company level. Unnesting \code{product} yields a data frame with at least columns \code{companies_id}, \code{grouped_by}, \code{risk_category}. Unnesting \code{company} yields a data frame with at least columns \code{companies_id}, \code{grouped_by}, \code{risk_category}, \code{value}. Any column in the input datasets ending with \verb{*rowid} is also passed as is to the output at product level. The exception is any column named exactly \code{rowid}-- which is a reserved name and throws an error. Note this feature makes no sense at company level because potentially multiple rows in the input datasets are summarized into a single row in the output at company level.
}
\description{
The indicator "sector profile" measures the transition risk of products based on
the sector's emissions targets the product belongs to. Those sector emission
reduction targets vary across scenarios (e.g., net zero scenario or 1.5°
scenario) and the time horizon (e.g., reduction needed in 2030, 2040, 2050 to
achieve the targets).

After assessing each product, all the products with the same category are
aggregated and set in relation to all products of the company. We, therefore,
derive company-level information.
}
\examples{
```{r child=extdata_path("child/example-sector_profile.Rmd")}
```
}
\seealso{
Other main functions: 
\code{\link{emissions_profile_upstream}()},
\code{\link{emissions_profile}()},
\code{\link{sector_profile_upstream}()}
}
\concept{main functions}
