% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/sector_profile_any_pivot_type_sector_subsector.R
\name{sector_profile_any_pivot_type_sector_subsector}
\alias{sector_profile_any_pivot_type_sector_subsector}
\title{Restructure XSTR \code{companies}}
\usage{
sector_profile_any_pivot_type_sector_subsector(data)
}
\arguments{
\item{data}{A dataframe with these columns:
\itemize{
\item ipr_sector
\item ipr_subsector
\item weo_product
\item weo_flow
}}
}
\value{
A \code{companies} dataset as required by sector functions.
}
\description{
Restructure XSTR \code{companies}
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(readr, warn.conflicts = FALSE)

raw_companies <- read_csv(extdata_path("pstr_companies.csv"))
glimpse(raw_companies)

companies <- xstr_pivot_type_sector_subsector(raw_companies)
companies
}
\seealso{
Other pre-processing helpers: 
\code{\link{sector_profile_any_prepare_scenario}()},
\code{\link{sector_profile_any_prune_companies}()}
}
\concept{pre-processing helpers}
