% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sector_profile_any_prune_companies.R
\name{sector_profile_any_prune_companies}
\alias{sector_profile_any_prune_companies}
\title{Drop rows where the product info is \code{NA} & sector info is duplicated}
\usage{
sector_profile_any_prune_companies(data)
}
\arguments{
\item{data}{Typically a "sector profile" \verb{*companies} dataframe.}
}
\value{
A dataframe with maybe fewer rows than the input \code{data}.
}
\description{
For each company, this function drops rows where the product information is
missing and the sector information is duplicated.
}
\examples{
library(dplyr)
# styler: off
companies <- tribble(
  ~row, ~company_id, ~clustered, ~activity_uuid_product_uuid, ~tilt_sector,
    1L,         "a",       "b1",                        "c1",          "x",
    2L,         "a",         NA,                          NA,          "x",
    3L,         "a",         NA,                          NA,          "y",
    4L,         "a",         NA,                          NA,          "y"
  )
# styler: off

# Keep row 1: Has product info
# Drop row 2: Lacks product info and sector info is duplicated
# Keep row 3: Lacks product info but sector info is unique
# Drop row 4: Lacks product info and sector info is duplicated
companies

sector_profile_any_prune_companies(companies)

}
\seealso{
Other pre-processing helpers: 
\code{\link{emissions_profile_any_compute_profile_ranking}()},
\code{\link{sector_profile_any_pivot_type_sector_subsector}()},
\code{\link{sector_profile_any_prepare_scenario}()}
}
\concept{pre-processing helpers}
