% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sector_profile_upstream.R
\name{sector_profile_upstream}
\alias{sector_profile_upstream}
\title{Calculate the indicator "sector profile upstream"}
\usage{
sector_profile_upstream(
  companies,
  scenarios,
  inputs,
  low_threshold = ifelse(scenarios$year == 2030, 1/9, 1/3),
  high_threshold = ifelse(scenarios$year == 2030, 2/9, 2/3)
)
}
\arguments{
\item{companies}{A dataframe like \link{istr_companies}.}

\item{scenarios}{A dataframe like \link{xstr_scenarios}.}

\item{inputs}{A dataframe like \link{istr_inputs}.}

\item{low_threshold}{A numeric value to segment low and medium transition
risk products.}

\item{high_threshold}{A numeric value to segment medium and high transition
risk products.}
}
\value{
A data frame with the column \code{companies_id}, and the nested columns \code{product} and \code{company} holding the outputs at product and company level. Unnesting \code{product} yields a data frame with at least columns \code{companies_id}, \code{grouped_by}, \code{risk_category}. Unnesting \code{company} yields a data frame with at least columns \code{companies_id}, \code{grouped_by}, \code{risk_category}, \code{value}.
}
\description{
\if{html}{\out{<div class="sourceCode {r child=extdata_path("child/intro-secor-profile-upstream.Rmd")}">}}\preformatted{}\if{html}{\out{</div>}}
}
\examples{
library(tiltToyData)
library(readr)
options(readr.show_col_types = FALSE)

toy_files()

companies <- read_csv(toy_path("sector_profile_upstream_companies.csv.gz"))
scenarios <- read_csv(toy_path("sector_profile_any_scenarios.csv.gz"))
inputs <- read_csv(toy_path("sector_profile_upstream_products.csv.gz"))

both <- sector_profile_upstream(companies, scenarios, inputs)
both

both |> unnest_product()

both |> unnest_company()
}
\concept{ISTR functions}
