% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sector_profile_upstream.R
\name{sector_profile_upstream}
\alias{sector_profile_upstream}
\title{Calculate the indicator "sector profile upstream"}
\usage{
sector_profile_upstream(
  companies,
  scenarios,
  inputs,
  low_threshold = ifelse(scenarios$year == 2030, 1/9, 1/3),
  high_threshold = ifelse(scenarios$year == 2030, 2/9, 2/3)
)
}
\arguments{
\item{companies, scenarios, inputs}{A dataframe like the dataset with a matching name in tiltToyData (see \href{https://2degreesinvesting.github.io/tiltToyData/reference/index.html}{Reference}).}

\item{low_threshold}{A numeric value to segment low and medium transition
risk products.}

\item{high_threshold}{A numeric value to segment medium and high transition
risk products.}
}
\value{
A data frame with the column \code{companies_id}, and the nested columns\code{product} and \code{company} holding the outputs at product and company level. Unnesting \code{product} yields a data frame with at least columns \code{companies_id}, \code{grouped_by}, \code{risk_category}. Unnesting \code{company} yields a data frame with at least columns \code{companies_id}, \code{grouped_by}, \code{risk_category}, \code{value}. Any column in the input datasets ending with \verb{*rowid} is also passed as is to the output at product level. The exception is any column named exactly \code{rowid}-- which is a reserved name and throws an error. Note this feature makes no sense at company level because potentially multiple rows in the input datasets are summarized into a single row in the output at company level.
}
\description{
The indicator "sector profile upstream" assesses the transition risk of the
input products based on the sector's emissions targets the input product belongs
to. This indicator can be aggregated on company level and as such inform about
the supply chain risk of an SME, based on its inputs' transition risk. The
sector emission reduction targets vary across scenarios (e.g., net zero scenario
or 1.5° scenario) and the time horizon (e.g., reduction needed in 2030, 2040,
2050 to achieve the targets). It, therefore, is similar to the Product Sector
Risk Indicator but focuses on the input products that the company needs to
produce its products.The input products are, for example, resources, packaging
materials, energy and enabling services (such as tractor use on farm) to produce
the product.

After identifying each carbon footprint for one input product, the input
products are ranked according to their footprint. The ranking method is explained in the
\href{https://2degreesinvesting.github.io/tiltIndicator/articles/tiltIndicator.html#thresholds}{Thresholds} section.

After assessing the input products for each product, they are aggregated at
company-level to derive what percentage of the input products required by the
company to produce its products have high, medium and low sector transition
risk. We, therefore, derive the company-level information.

Please note that carbon emissions or emissions always mean CO2e.
}
\examples{
library(tiltIndicator)
library(tiltToyData)
library(readr)
options(readr.show_col_types = FALSE)

companies <- read_csv(toy_sector_profile_upstream_companies())
scenarios <- read_csv(toy_sector_profile_any_scenarios())
inputs <- read_csv(toy_sector_profile_upstream_products())

both <- sector_profile_upstream(companies, scenarios, inputs)
both

both |> unnest_product()

both |> unnest_company()
}
\seealso{
Other main functions: 
\code{\link{emissions_profile_upstream}()},
\code{\link{emissions_profile}()},
\code{\link{sector_profile}()}
}
\concept{main functions}
