% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/istr_at_product_level.R,
%   R/sector_profile_upstream.R, R/xctr.R
\name{istr_at_product_level}
\alias{istr_at_product_level}
\alias{sector_profile_upstream}
\alias{istr_at_company_level}
\title{Calculate the ISTR indicator}
\usage{
istr_at_product_level(
  companies,
  scenarios,
  inputs,
  low_threshold = ifelse(scenarios$year == 2030, 1/9, 1/3),
  high_threshold = ifelse(scenarios$year == 2030, 2/9, 2/3)
)

sector_profile_upstream(
  companies,
  scenarios,
  inputs,
  low_threshold = ifelse(scenarios$year == 2030, 1/9, 1/3),
  high_threshold = ifelse(scenarios$year == 2030, 2/9, 2/3)
)

istr_at_company_level(data)
}
\arguments{
\item{companies}{A dataframe like \link{istr_companies}.}

\item{scenarios}{A dataframe like \link{xstr_scenarios}.}

\item{inputs}{A dataframe like \link{istr_inputs}.}

\item{low_threshold}{A numeric value to segment low and medium transition
risk products.}

\item{high_threshold}{A numeric value to segment medium and high transition
risk products.}

\item{data}{A dataframe. The output at product level.}
}
\value{
A data frame with the column \code{companies_id}, and the nested columns \code{product} and \code{company} holding the outputs at product and company level. Unnesting \code{product} yields a data frame with at least columns \code{companies_id}, \code{grouped_by}, \code{risk_category}. Unnesting \code{company} yields a data frame with at least columns \code{companies_id}, \code{grouped_by}, \code{risk_category}, \code{value}.
}
\description{
The Input Product Sector Risk Indicator assesses the transition risk of the
input products based on the sector's emissions targets the input product belongs
to. This indicator can be aggregated on company level and as such inform about
the supply chain risk of an SME, based on its inputs’ transition risk. The
sector emission reduction targets vary across scenarios (e.g., net zero scenario
or 1.5° scenario) and the time horizon (e.g., reduction needed in 2030, 2040,
2050 to achieve the targets). It, therefore, is similar to the Product Sector
Risk Indicator but focuses on the input products that the company needs to
produce its products.The input products are, for example, resources, packaging
materials, energy and enabling services (such as tractor use on farm) to produce
the product.

After identifying each carbon footprint for one input product, the input
products are ranked according to their footprint. The ranking method is explained in the
\href{https://2degreesinvesting.github.io/tiltIndicator/articles/tiltIndicator.html#thresholds}{Thresholds} section.

After assessing the input products for each product, they are aggregated at
company-level to derive what percentage of the input products required by the
company to produce its products have high, medium and low sector transition
risk. We, therefore, derive the company-level information.

Please note that carbon emissions or emissions always mean CO2e.
}
\examples{
companies <- istr_companies
scenarios <- xstr_scenarios
inputs <- istr_inputs

# Product level
companies |>
  istr_at_product_level(scenarios, inputs)

# Company level
companies |>
  istr_at_product_level(scenarios, inputs) |>
  istr_at_company_level()

# Same
both <- istr(companies, scenarios, inputs)
both

both |> unnest_product()

both |> unnest_company()
}
\concept{ISTR functions}
