% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest.R
\name{unnest_product}
\alias{unnest_product}
\alias{unnest_company}
\title{Unnest product- and company-level results}
\usage{
unnest_product(data)

unnest_company(data)
}
\arguments{
\item{data}{A nested data frame, e.g. the output of \code{\link[=sector_profile]{sector_profile()}}.}
}
\value{
A data frame.
}
\description{
Unnest product- and company-level results
}
\examples{
library(tiltToyData)
library(readr)
options(readr.show_col_types = FALSE)

toy_files()

companies <- read_csv(toy_path("sector_profile_companies.csv.gz"))
scenarios <- read_csv(toy_path("sector_profile_any_scenarios.csv.gz"))

both <- sector_profile(companies, scenarios)
both

both |> unnest_product()

both |> unnest_company()
}
\concept{helpers}
