% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_refactoring_checklist.R
\name{use_refactoring_checklist}
\alias{use_refactoring_checklist}
\title{Create a task list that you can paste on a GitHub issue and follow up}
\usage{
use_refactoring_checklist(indicator)
}
\arguments{
\item{indicator}{Character. The indicator abbreviation, e.g. for
"product-carbon-transition-risk" use "pstr".}
}
\value{
Character. A checklist you can paste on a GitHub issue.
}
\description{
Create a task list that you can paste on a GitHub issue and follow up
}
\examples{
use_refactoring_checklist("pstr")
}
\seealso{
Other internals: 
\code{\link{mvp_path}()},
\code{\link{render_to_list}()}
}
\concept{internals}
