% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstr_data.R
\docType{data}
\name{weo_2022}
\alias{weo_2022}
\title{A dataset with scenario, year, product and reductions information}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8 rows and 10 columns.
}
\usage{
weo_2022
}
\description{
\itemize{
\item TODO: Where does it come from and what wrangling did you do?
}
}
\examples{
weo_2022
}
\seealso{
Other datasets for product sector transition risk.: 
\code{\link{companies}},
\code{\link{ep_weo}}
}
\concept{datasets for product sector transition risk.}
\keyword{datasets}
