% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xstr_polish_output_at_copmany_level.R
\name{xstr_polish_output_at_company_level}
\alias{xstr_polish_output_at_company_level}
\title{Polish output at company level}
\usage{
xstr_polish_output_at_company_level(data)
}
\arguments{
\item{data}{The output of \verb{xstr*()} functions.}
}
\value{
A dataframe.
}
\description{
Polish output at company level
}
\examples{
library(tiltToyData)
library(readr)
options(readr.show_col_types = FALSE)

toy_files()

companies <- toy_path("sector_profile_companies.csv.gz") |>
  read_csv()
scenarios <- toy_path("sector_profile_any_scenarios.csv.gz") |>
  read_csv()
sector_profile(companies, scenarios) |>
  unnest_company() |>
  xstr_polish_output_at_company_level()

companies <- toy_path("sector_profile_upstream_companies.csv.gz") |>
  read_csv()
upstream_products <- toy_path("sector_profile_upstream_products.csv.gz") |>
  read_csv(col_types = cols(input_isic_4digit = col_character()))
sector_profile_upstream(companies, scenarios, upstream_products) |>
  unnest_company() |>
  xstr_polish_output_at_company_level()
}
\concept{post-processing helpers}
