% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xstr_polish_output_at_copmany_level.R
\name{xstr_polish_output_at_company_level}
\alias{xstr_polish_output_at_company_level}
\title{Polish output at company level}
\usage{
xstr_polish_output_at_company_level(data)
}
\arguments{
\item{data}{The output of \verb{xstr*()} functions.}
}
\value{
A dataframe.
}
\description{
Polish output at company level
}
\examples{
pstr(pstr_companies, pstr_scenarios) |>
  xstr_polish_output_at_company_level()

istr(istr_companies, istr_scenarios, istr_inputs) |>
  xstr_polish_output_at_company_level()
}
\concept{post-processing helpers}
