% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xstr_prune_companies.R
\name{xstr_prune_companies}
\alias{xstr_prune_companies}
\title{Drop rows where the product info is \code{NA} & sector info is duplicated}
\usage{
xstr_prune_companies(data)
}
\arguments{
\item{data}{Typically an XSTR \verb{*companies} dataframe.}
}
\value{
A dataframe with maybe fewer rows than the input \code{data}.
}
\description{
For each company, this function drops rows where the product information is
missing and the sector information is duplicated.
}
\examples{
# styler: off
companies <- tibble::tribble(
  ~row_id, ~companies_id, ~clustered, ~activity_uuid_product_uuid, ~tilt_sector,
  # Keep: Has product info
        1,           "a",       "b1",                        "c1",          "x",
  # Drop: Lacks product info and sector info is duplicated
        2,           "a",         NA,                          NA,          "x",
  # Drop: Lacks product info and sector info is duplicated
        3,           "a",         NA,                          NA,          "x",
  # Keep: Lacks product info but sector info is unique
        4,           "a",         NA,                          NA,          "y",
  # Drop: Lacks product info and sector info is duplicated
        5,           "a",         NA,                          NA,          "y",
  # Keep: Has product info
        6,           "a",       "b6",                        "c6",          "z",
  # Keep: Has product info
        7,           "a",       "b7",                          NA,          "z",
)
# styler: on

xstr_prune_companies(companies)
}
\seealso{
Other helpers: 
\code{\link{pstr_polish_output_at_copmany_level}()},
\code{\link{pstr_prepare_scenario}()},
\code{\link{xstr_pivot_type_sector_subsector}()}
}
\concept{helpers}
